### 2025-09-02 #8911 @crvernon
:wave: @matbun  - Thank you for your submission to JOSS.  All of our topic editors are occupied currently as we are managing a large queue of submissions.  I will get you a topic editor assigned as soon as possible.

In the meantime, could you please reduce the word count of your paper to <= 1000 words.  Feel free to reference content from your documentation in your paper to reduce this length.

Thank you.
