### 2025-09-04 0802fc4
First attempt to reduce wc
