+++ joss/paper.md
@@ -79,62 +79,41 @@ bibliography: paper.bib
 
 # Summary
 
-The integration of Artificial Intelligence (AI) into scientific research has
-expanded significantly over the past decade, driven by the availability of
-large-scale datasets and specialized hardware accelerators. Graphics Processing
-Units (GPUs) have become essential for AI workloads, with modern
-High-Performance Computing (HPC) facilities housing thousands of accelerators:
-the JUWELS Booster system [@JUWELS] at the Jülich Supercomputing Center (JSC),
-Forschungszentrum Jülich, operates over 3,700 NVIDIA A100 GPUs, while Finland's
-LUMI supercomputer provides access to more than 11,000 AMD MI250X GPUs, and the
-upcoming first European Exascale system JUPITER hosted at JSC, expanding capabilities
-up to 24,000 NVIDIA GH200 GPUs.
+Artificial Intelligence (AI) is increasingly used in scientific research,
+supported by large datasets and HPC systems equipped with modern hardware
+accelerators.  Facilities such as [JUWELS
+Booster](https://apps.fz-juelich.de/jsc/hps/juwels/booster-overview.html),
+[LUMI](https://www.lumi-supercomputer.eu/), and the upcoming
+[JUPITER](https://www.fz-juelich.de/en/ias/jsc/jupiter) system enable
+large-scale training across scientific domains.
 
-This growth has enabled development of numerous domain-specific AI solutions
-across scientific fields, offering the potential for faster experimentation and
-shorter feedback cycles. However, many researchers face significant barriers
-when deploying AI workflows on HPC systems due to the complexity of, e.g.,
-running distributed training and Hyperparameter Optimization (HPO) at scale, or
-in utilizing accelerator technologies efficiently. The heterogeneous nature of
-HPC systems forces scientists to focus on low-level implementation details
-rather than on their core research.
+Despite these advancements, researchers often face difficulties in deploying AI
+workflows on HPC. Challenges include the complexity of distributed training,
+hyperparameter optimization, and adapting to heterogeneous architectures. These
+issues divert focus from scientific goals to technical details.
 
-To address these challenges, we present `itwinai`, a Python library designed to
-bridge the gap between domain sciences, AI, and HPC. The library provides a
-standardized interface, enabling seamless scaling from laptops to
-supercomputers.  Through its modular architecture, `itwinai` significantly
-lowers the entry barrier for domain specialists while ensuring optimal HPC
-resource utilization.
+To address this, we present `itwinai`, a Python library that simplifies scalable
+AI on HPC. Its modular architecture and standard interface allow users to scale
+workloads efficiently from laptops to supercomputers, reducing implementation
+overhead and improving resource usage.
 
 # Statement of need
 
-The integration of diverse ML tools and libraries requires significant
-integration effort on the part of the scientific users, especially with HPC
-systems. `itwinai` is a Python library that streamlines the integration of
-AI-powered scientific workflows into HPC infrastructure through ready-to-use
-components. The library addresses three core scientific Machine Learning (ML)
-needs:
+Integrating machine learning into scientific workflows on HPC systems remains
+complex. Researchers must often invest substantial effort to configure
+distributed training, manage hyperparameter optimization, and analyze
+performance, while adapting to varied system architectures.
 
-**Distributed training** with uniform interfaces across
-PyTorch-DDP [@torch_ddp], DeepSpeed [@deepspeed], and Horovod [@horovod]
-backends.
+`itwinai` is a Python library that streamlines this process by providing a
+unified framework for scalable AI workflows. It offers consistent interfaces for
+distributed training, supports large-scale hyperparameter optimization, and
[…]
