+++ joss/compile.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 
 docker run --rm -it \
-    -v $PWD:/data \
-    -u $(id -u):$(id -g) \
+    -v "$PWD:/data" \
+    -u "$(id -u):$(id -g)" \
     openjournals/inara \
     -o pdf,crossref \
     paper.md
