### 2025-09-08 76d0e6e
Further reduction of word count
