+++ joss/paper.md
@@ -81,39 +81,28 @@ bibliography: paper.bib
 
 The integration of Artificial Intelligence (AI) into scientific research has
 expanded significantly over the past decade, driven by the availability of
-large-scale datasets and specialized hardware accelerators. Graphics Processing
-Units (GPUs) have become essential for AI workloads, with modern
-High-Performance Computing (HPC) facilities housing thousands of accelerators:
-the JUWELS Booster system [@JUWELS] at the Jülich Supercomputing Center (JSC),
-Forschungszentrum Jülich, operates over 3,700 NVIDIA A100 GPUs, while Finland's
-LUMI supercomputer provides access to more than 11,000 AMD MI250X GPUs, and the
-upcoming first European Exascale system JUPITER hosted at JSC, expanding capabilities
-up to 24,000 NVIDIA GH200 GPUs.
+large-scale datasets and graphic processing units (GPUs),
+in particular at high performance computing (HPC) sites.
 
-This growth has enabled development of numerous domain-specific AI solutions
-across scientific fields, offering the potential for faster experimentation and
-shorter feedback cycles. However, many researchers face significant barriers
-when deploying AI workflows on HPC systems due to the complexity of, e.g.,
-running distributed training and Hyperparameter Optimization (HPO) at scale, or
-in utilizing accelerator technologies efficiently. The heterogeneous nature of
-HPC systems forces scientists to focus on low-level implementation details
-rather than on their core research.
+However, many researchers face significant barriers
+when deploying AI workflows on HPC systems, as their heterogeneous nature of
+forces scientists to focus on low-level implementation details
+rather than on their core research. At the same time, the researchers often lack 
+specialized HPC/AI knowledge to implement their workflows efficiently. 
 
 To address these challenges, we present `itwinai`, a Python library designed to
-bridge the gap between domain sciences, AI, and HPC. The library provides a
-standardized interface, enabling seamless scaling from laptops to
-supercomputers.  Through its modular architecture, `itwinai` significantly
+bridge the gap between domain sciences, AI, and HPC. 
+Through its modular architecture, `itwinai` significantly
 lowers the entry barrier for domain specialists while ensuring optimal HPC
 resource utilization.
 
 # Statement of need
 
-The integration of diverse ML tools and libraries requires significant
+The integration of ML tools requires significant
 integration effort on the part of the scientific users, especially with HPC
 systems. `itwinai` is a Python library that streamlines the integration of
-AI-powered scientific workflows into HPC infrastructure through ready-to-use
-components. The library addresses three core scientific Machine Learning (ML)
-needs:
+AI-powered scientific workflows into HPC infrastructure by addressing
+three core scientific Machine Learning (ML) needs:
 
 **Distributed training** with uniform interfaces across
 PyTorch-DDP [@torch_ddp], DeepSpeed [@deepspeed], and Horovod [@horovod]
@@ -124,14 +113,11 @@ backends.
 **Comprehensive performance analysis** including profiling, scaling
 and parallel efficiency metrics, and bottleneck identification.
 
-Built with a modular architecture, `itwinai` enables component integration and
-plugin-based extensions for independent development of scientific use cases.
+Built with a modular architecture, `itwinai` uses
+plugin-based extensions for independent application to scientific use cases.
 Developed as part of the interTwin[^intertwin_eu] project to support Digital Twin
 applications across physics and environmental sciences, the library has proven
-versatile for general AI applications. By consolidating diverse functionalities
-into a single framework, `itwinai` significantly lowers the barrier to HPC
[…]
