### 2025-09-08 7821003
further small changes to reduce word count
