+++ joss/paper.md
@@ -79,18 +79,14 @@ bibliography: paper.bib
 
 # Summary
 
-Artificial Intelligence (AI) is increasingly used in scientific research,
-supported by large datasets and HPC systems equipped with modern hardware
-accelerators.  Facilities such as [JUWELS
-Booster](https://apps.fz-juelich.de/jsc/hps/juwels/booster-overview.html),
-[LUMI](https://www.lumi-supercomputer.eu/), and the upcoming
-[JUPITER](https://www.fz-juelich.de/en/ias/jsc/jupiter) system enable
-large-scale training across scientific domains.
+The integration of Artificial Intelligence (AI) into scientific research has
+expanded significantly over the past decade, driven by the availability of
+large-scale datasets and graphic processing units (GPUs),
+in particular at high performance computing (HPC) sites.
 
-Despite these advancements, researchers often face difficulties in deploying AI
-workflows on HPC. Challenges include the complexity of distributed training,
-hyperparameter optimization, and adapting to heterogeneous architectures. These
-issues divert focus from scientific goals to technical details.
+However, many researchers face significant barriers
+when deploying AI workflows on HPC systems, as their heterogeneous nature forces scientists to focus on low-level implementation details
+rather than on their core research. At the same time, the researchers often lack specialized HPC/AI knowledge to implement their workflows efficiently. 
 
 To address this, we present `itwinai`, a Python library that simplifies scalable
 AI on HPC. Its modular architecture and standard interface allow users to scale
@@ -99,48 +95,50 @@ overhead and improving resource usage.
 
 # Statement of need
 
-Integrating machine learning into scientific workflows on HPC systems remains
-complex. Researchers must often invest substantial effort to configure
-distributed training, manage hyperparameter optimization, and analyze
-performance, while adapting to varied system architectures.
+The integration of ML tools requires significant
+integration effort on the part of the scientific users, especially with HPC
+systems. `itwinai` is a Python library that streamlines the integration of
+AI-powered scientific workflows into HPC infrastructure by addressing
+three core scientific Machine Learning (ML) needs:
 
-`itwinai` is a Python library that streamlines this process by providing a
-unified framework for scalable AI workflows. It offers consistent interfaces for
-distributed training, supports large-scale hyperparameter optimization, and
-includes tools for profiling and code scalability analysis.
+**Distributed training** with uniform interfaces across
+PyTorch-DDP [@torch_ddp], DeepSpeed [@deepspeed], and Horovod [@horovod]
+backends.
 
-Developed within the [interTwin](https://www.intertwin.eu/) project to support Digital Twin
-applications in physics and environmental sciences, `itwinai` is designed to be
-extensible and reusable. By consolidating core functionality into a single
-framework, it lowers the technical barrier to HPC adoption and enables
-researchers to focus on scientific objectives.
+**Scalable HPO** via Ray Tune [@ray_tune] supporting various search algorithms.
+
+**Comprehensive performance analysis** including profiling, scaling
+and parallel efficiency metrics, and bottleneck identification.
+
+Built with a modular architecture, `itwinai` uses
+plugin-based extensions for independent application to scientific use cases.
+Developed as part of the interTwin[^intertwin_eu] project to support Digital Twin
+applications across physics and environmental sciences, the library has proven
+versatile for general AI applications.
+
+[^intertwin_eu]: interTwin project: [intertwin.eu](https://www.intertwin.eu/)
+(Accessed on 2025-08-14).
 
 # Package features
 
 The main features offered by the `itwinai` library are:
 
[…]
