### 2025-09-11 773fe89
Reducing WC for JOSS (#458)
* First attempt to reduce wc

* fix linter checks

* further small changes to reduce word count

* Capitalisations and removing commented block

* fixing linter errors

* linter

---------

Co-authored-by: Alex Krochak <o.krochak@fz-juelich.de>
Co-authored-by: Rakesh Sarma <126173968+r-sarma@users.noreply.github.com>
Co-authored-by: Alex <78349687+okrochak@users.noreply.github.com>
