+++ joss/paper.md
@@ -191,12 +191,15 @@ components are provided: scalability report generation and profiling.
 
 ## Scalability report
 
-Wall-clock epoch time, relative speedup
+For data-parallel training, adding more workers improves throughput, but
+as all-reduce communication costs grow, communication overhead eventually dominates,
+causing scaling to level-off or even decline. The report characterizes this trade-off
+across GPUs/nodes and backends, reporting wall-clock epoch time, relative speedup
 (\autoref{fig:speedup}), GPU utilization (0–100%), energy (Wh), and
-compute-versus-other time are provided (\autoref{fig:compvsother}).
-
-Considered jointly, these metrics
-identify the most efficient configuration and distribution strategy. \autoref{fig:speedup} and
+compute-versus-other time, including collective communication and memory
+operations (\autoref{fig:compvsother}). Considered jointly, these metrics
+identify the most efficient configuration and distribution strategy, rather than
+relying on a single indicator. \autoref{fig:speedup} and
 \autoref{fig:compvsother} show the scalability of the physics use case from
 INFN[^infn] targeting gravitational-wave analysis at the Virgo[^virgo]
 interferometer [@tsolaki_2025_15120028] [@saether_scalabiliy_2025].
