### 2025-09-12 694a68f
remove some footnotes and simplify statement of need
