### 2025-11-13 #8911 @matbun
> Hi [@matbun](https://github.com/matbun)
>
> A few thoughts currently:
>
> * Add a short section “State of the field” comparing with some other repos. Spell out what itwinai uniquely offers (HPC/SLURM portability, plugin system, HPO/profiling/logging, cloud↔HPC offloading) and when a user should pick it. Also taking a look at [Checklist](https://joss.readthedocs.io/en/latest/review_checklist.html) would be helpful.
> * Reproducibility for figures: please list more to rerun/compare: hardware (GPU/CPU, nodes, interconnect, driver/SDK), software versions + backend settings, workload (dataset/model, batch size, epochs, optimizer), and exact commands/config paths.
> * Minor nits “suppported” → “supported”;

Hi @yewentao256

Thanks a lot for your feedback.

Unfortunately we already had to cut a lot from our original version to get closer to the 1k words limit. We are still a bit above that limit, and I am afraid that if we add more explanations we will have to cut text from other sections, which may reduce their clarity. How strict is the 1k words cap? If this is not too strict, we could easily address your point by adding a couple hundreds more words.

Regarding reproducibility we could push code, configuration files, and additional information to Zenodo (example: https://zenodo.org/records/16882390) and reference it from the paper. This would help us reducing the number of words to add.
Would this be a reasonable solution?

Thank you.
