### 2025-11-13 #8911 @yewentao256
Don't worry too much about the words limit, `https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain`

>We also require that authors explain the research applications of the software. The paper should be between 250-1000 words. Authors submitting papers **significantly longer than 1000 words** `may` be asked to reduce the length of their paper.

1300 words still look good on my side.

For the `reproducibility`, I think you can:
1. Add more instructions and related set up in detail in your repo README
2. Briefly talk about the config in your paper, like hardware (GPU etc).
