### 2025-11-14 #8911 @boisgera
Hi @yewentao256, @matbun,

@xiazeyu joined us as a reviewer!

Therefore, I am about to officially open the review issue, which will automatically close this pre-review issue.

Please continue your discussion in [the JOSS review issue](https://github.com/openjournals/joss-reviews/issues/9409) (or in the [issues of the project](https://github.com/interTwin-eu/itwinai/issues) when it's more appropriate).

Thanks everyone!
