### 2025-11-14 #9409 @editorialbot
[REVIEW]: itwinai: A Python Toolkit for Scalable Scientific Machine Learning on HPC Systems
**Submitting author:** @matbun (<a href="http://orcid.org/0009-0008-5100-9300">Matteo Bunino</a>)
**Repository:** https://github.com/interTwin-eu/itwinai
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.4.1
