+++ .github/workflows/check-links.yml
@@ -13,7 +13,7 @@ jobs:
     steps:
       # Checks out a copy of your repository on the ubuntu-latest machine
       - name: Checkout code
-        uses: actions/checkout@v5
+        uses: actions/checkout@v6
         with:
           # Make sure the actual branch is checked out when running on PR
           # ref: ${{ github.event.pull_request.head.sha }}
