### 2025-12-01 96546ab
Fix typos and improve clarity in paper.md
Corrected minor typographical errors and improved clarity in the document.
