### 2025-12-01 b2a9589
Update references in paper.md for clarity
