### 2025-12-01 79903dd
Fix typos and improve clarity in paper.md (#462)
* Fix typos and improve clarity in paper.md

Corrected minor typographical errors and improved clarity in the document.

* Update references in paper.md for clarity
