### 2025-12-01 interTwin-eu/itwinai#463 @matbun
[JOSS] State of the field section
Hi @r-sarma, @okrochak, could you help me address [this comment](https://github.com/openjournals/joss-reviews/issues/8911#issuecomment-3505189252) from one of our JOSS reviewers on related work?

> Add a short section “State of the field” comparing with some other repos. Spell out what itwinai uniquely offers (HPC/SLURM portability, plugin system, HPO/profiling/logging, cloud↔HPC offloading) and when a user should pick it. Also taking a look at [Checklist](https://joss.readthedocs.io/en/latest/review_checklist.html) would be helpful.

Other projects I can think of are:
- [perun](https://github.com/Helmholtz-AI-Energy/perun)
- [heat](https://github.com/helmholtz-analytics/heat/)
- [propulate](https://github.com/Helmholtz-AI-Energy/propulate)

Do you think these projects are worth to be mentioned? Do you know others?

Thanks!
