### 2025-12-02 interTwin-eu/itwinai#464 @matbun
[JOSS] Resources for plots reproducibility
Address the requests from Ye Wentao here: https://github.com/openjournals/joss-reviews/issues/9409#issuecomment-3599390397
