### 2025-12-02 a755289
Add 'State of the field' section to paper
Added a section on the state of the field discussing scalable AI workflows on HPC systems and the capabilities of the `itwinai` library.
