### 2025-12-02 b2614ec
Add new references for scalability and profiling
