### 2025-12-02 add10ce
Add reproducibility section to paper
Added a section on reproducibility detailing the scalability results obtained on the JUWELS Booster GPU partition, including configuration file management and reproducibility of results.
