### 2025-12-02 d3e5859
Update URLs in bibliography for versioning
