+++ joss/paper.bib
@@ -240,7 +240,7 @@ This data was used to generate the plots in the JOSS submission for itwinai},
   year      = {2025},
   publisher = {Read the Docs},
   journal   = {itwinai documentation},
-  url       = {https://itwinai.readthedocs.io/latest/how-it-works/scalability-report/scalability_report.html},
+  url       = {https://itwinai.readthedocs.io/v0.3.4/how-it-works/scalability-report/scalability_report.html},
   urldate   = {2025-08-14},
 }
 
@@ -250,6 +250,6 @@ This data was used to generate the plots in the JOSS submission for itwinai},
   year      = {2025},
   publisher = {Read the Docs},
   journal   = {itwinai documentation},
-  url       = {https://itwinai.readthedocs.io/latest/tutorials/profiling/profiling-overview.html},
+  url       = {https://itwinai.readthedocs.io/v0.3.4/tutorials/profiling/profiling-overview.html},
   urldate   = {2025-08-14},
 }
