### 2025-12-03 bd1bf11
Update paper.md
