### 2025-12-03 d971392
Add state of the field (#465)
* Add new bibliographic entries to paper.bib

Added multiple references including Optuna, HeAT, AI4HPC, and Perun.

* Add 'State of the field' section to paper

Added a section on the state of the field discussing scalable AI workflows on HPC systems and the capabilities of the `itwinai` library.

* Update paper.md

---------

Co-authored-by: Rakesh Sarma <126173968+r-sarma@users.noreply.github.com>
