+++ joss/paper.bib
@@ -31,6 +31,47 @@
 	year = {2025},
 }
 
+@inproceedings{optuna_2019,
+  author    = {Akiba, Takuya and Sano, Shotaro and Yanase, Toshihiko and Ohta, Takeru and Koyama, Masanori},
+  title     = {Optuna: A Next-generation Hyperparameter Optimization Framework},
+  booktitle = {Proceedings of the 25th {ACM} {SIGKDD} International Conference on Knowledge Discovery and Data Mining},
+  year      = {2019},
+  pages     = {2623--2631},
+  publisher = {ACM},
+  doi       = {10.1145/3292500.3330701},
+  url       = {https://doi.org/10.1145/3292500.3330701},
+}
+
+@misc{heat,
+  author    = {HeAT Developers},
+  title     = {HeAT},
+  year      = {2025},
+  publisher = {GitHub},
+  journal   = {GitHub repository},
+  url       = {https://github.com/helmholtz-analytics/heat},
+  urldate   = {2025-08-14},
+}
+
+@misc{ai4hpc,
+  author    = {AI4HPC Developers},
+  title     = {AI4HPC},
+  year      = {2025},
+  publisher = {GitLab},
+  journal   = {GitLab repository},
+  url       = {https://gitlab.jsc.fz-juelich.de/CoE-RAISE/FZJ/ai4hpc/ai4hpc},
+  urldate   = {2025-08-14},
+}
+
+@misc{perun,
+  author    = {Helmholtz AI Energy Team},
+  title     = {perun},
+  year      = {2025},
+  publisher = {GitHub},
+  journal   = {GitHub repository},
+  url       = {https://github.com/Helmholtz-AI-Energy/perun},
+  urldate   = {2025-08-14},
+}
+
 @article{benach_intertwin_2023,
 	title = {{interTwin} {D3}.1 {DTE} blueprint architecture, functional specifications and requirements analysis first version},
 	url = {https://zenodo.org/records/10417124},
