### 2025-12-03 e07a2f8
Draft Virgo code
