### 2025-12-03 9a49cb2
Code cleanup
