+++ joss/reproducibility/trainer.py
@@ -215,10 +215,10 @@ class NoiseGeneratorTrainer(TorchTrainer):
         val_acc_plot = []
         best_val_loss = float("inf")
 
-        for epoch in tqdm(range(self.num_epochs)):
+        for self.current_epoch in tqdm(range(self.num_epochs)):
             lt = timer()
             # itwinai - IMPORTANT: set current epoch ID
-            self.set_epoch(epoch)
+            self.set_epoch()
             t_list = []
             st = time.time()
             epoch_loss = []
@@ -248,7 +248,7 @@ class NoiseGeneratorTrainer(TorchTrainer):
                     loss.detach().cpu().numpy(),
                     "epoch_loss_batch",
                     kind="metric",
-                    step=epoch * len(self.train_dataloader) + i,
+                    step=self.current_epoch * len(self.train_dataloader) + i,
                     batch_idx=i,
                 )
                 # acc=accuracy(generated.detach().cpu().numpy(),target.detach().cpu().numpy(),20)
@@ -274,7 +274,7 @@ class NoiseGeneratorTrainer(TorchTrainer):
                     loss.detach().cpu().numpy(),
                     "val_loss_batch",
                     kind="metric",
-                    step=epoch * len(self.validation_dataloader) + i,
+                    step=self.current_epoch * len(self.validation_dataloader) + i,
                     batch_idx=i,
                 )
                 # acc=accuracy(generated.detach().cpu().numpy(),target.detach().cpu().numpy(),20)
@@ -285,8 +285,8 @@ class NoiseGeneratorTrainer(TorchTrainer):
             # val_acc_plot.append(np.mean(val_acc))
 
             # itwinai - Log metrics/losses
-            self.log(np.mean(epoch_loss), "epoch_loss", kind="metric", step=epoch)
-            self.log(np.mean(val_loss), "val_loss", kind="metric", step=epoch)
+            self.log(np.mean(epoch_loss), "epoch_loss", kind="metric", step=self.current_epoch)
+            self.log(np.mean(val_loss), "val_loss", kind="metric", step=self.current_epoch)
             # self.log(np.mean(epoch_acc), 'epoch_acc',
             #          kind='metric', step=epoch)
             # self.log(np.mean(val_acc), 'val_acc',
@@ -300,12 +300,12 @@ class NoiseGeneratorTrainer(TorchTrainer):
             if self.strategy.is_main_worker:
                 print(
                     "epoch: {} loss: {} val loss: {} time:{}s".format(
-                        epoch, loss_plot[-1], val_loss_plot[-1], et - st
+                        self.current_epoch, loss_plot[-1], val_loss_plot[-1], et - st
                     )
                 )
 
             # Save checkpoint every #validation_every epochs
-            if self.validation_every and epoch % self.validation_every == 0:
+            if self.validation_every and self.current_epoch % self.validation_every == 0:
                 # uncomment the following if you want to save checkpoint every
                 # 100 epochs regardless of the performance of the model
                 # checkpoint = {
@@ -334,7 +334,7 @@ class NoiseGeneratorTrainer(TorchTrainer):
                     if self.config.save_best and avg_loss < best_val_loss:
                         # create checkpoint
                         checkpoint = {
-                            "epoch": epoch,
+                            "epoch": self.current_epoch,
                             "model_state_dict": self.model.state_dict(),
                             "optim_state_dict": self.optimizer.state_dict(),
                             "loss": loss_plot[-1],
@@ -343,7 +343,7 @@ class NoiseGeneratorTrainer(TorchTrainer):
 
                         # save checkpoint only if it is better than
                         # the previous ones
-                        checkpoint_filename = self.checkpoints_location.format(epoch)
+                        checkpoint_filename = self.checkpoints_location.format(self.current_epoch)
[…]

[…]
