### 2025-12-03 55215ac
Add profiling metrics to pipeline
