### 2025-12-05 cd032cc
First version datagen tutorial
