+++ joss/reproducibility/README.md
@@ -18,30 +18,96 @@ single-host setting you have access to Docker or equivalent (e.g., podman).
 
 ## Environment
 
-Below some instruction on how to reuse the same dependencies and itwinai version.
+Below some instruction on how to reuse the same dependencies and itwinai
+version.
 
-The results in the paper were obtained with itwinai `v0.3.4`, and the same dependencies can be
-found by using the following OCI container:
+The results in the paper were obtained with itwinai `v0.3.4`, and the same
+dependencies can be found by using the following OCI container:
 `ghcr.io/intertwin-eu/itwinai:joss-virgo-experiments`
 
-HPC systems generally prefer Apptainer or Singularity over Docker. If on an HPC, first
-download the container:
+HPC systems generally prefer Apptainer or Singularity over Docker. If on an HPC,
+first download the container:
 
 ```bash
 singularity pull itwinai.sif docker://ghcr.io/intertwin-eu/itwinai:joss-virgo-experiments
 ```
 
+## Machine learning workflow
+
+In this case, for our experiments we will use a small synthetic dataset
+generated on the fly.  The original data collected at the Virgo detector is not
+open.
+
+The machine learning workflow is described in the `training_pipeline_small` in
+the `config.yaml` configuration file. This pipeline describes add the steps in
+the ML workflow, including dataset generation, train-validation split,
+pre-processing, and training.
+
+Alternatively, you could generate a larger dataset and run a more
+compute-intensive workload using the `training_pipeline` pipeline in
+`config.yaml`. However, this is not covered in this guide and the interested
+reader is referred to the `README-use-case.md` file.
+
 ## Dataset Generation
 
-First we need to generate a synthetic training dataset. For the sake of simplicity, we will
-generate a smaller one compared to the one used in the experiments reported in the paper.
-Therefore, the results may not be identical.
+First we need to generate a synthetic training dataset. For the sake of
+simplicity, we will generate a smaller one compared to the one used in the
+experiments reported in the paper.  Therefore, the results may not be identical.
 
-On an HPC system... TODO
-
-On an single-host setup,
+On an single-host setup, generate a sample training dataset by executing only
+the first step in the pipeline. Notice that the local `$PWD/data` directory is
+mounted on `/data` to store the data on a persistent location. Remember to mount
+that every time you want to use the generated dataset.
 
 ```bash
-docker run ghcr.io/intertwin-eu/itwinai:joss-virgo-experiments \
+docker run --rm -v "$PWD/data":/data ghcr.io/intertwin-eu/itwinai:joss-virgo-experiments \
+    itwinai exec-pipeline +pipe_key=training_pipeline_small +pipe_steps=[0] 
+```
+
+On an HPC system you can run the same command on a login node with small
+adaptations, or submit a SLURM job in the following way:
+
+```bash
+sbatch synthetic-data-gen/data_generation_small.sh
+```
+
+At the end of dataset generation, you should see some pickle files:
+
+```bash
+$ ls data/*.pkl
+Image_dataset_synthetic_64x64.pkl  TimeSeries_dataset_synthetic_aux.pkl  TimeSeries_dataset_synthetic_main.pkl
+```
+
+To force the re-generation of the dataset, remove them.
+
+## Training and scalability metrics
+
+To begin with, please refer to the [scalability report
+guide](https://itwinai.readthedocs.io/v0.3.4/how-it-works/scalability-report/scalability_report.html)
+and to the [introduction to profiling with
+itwinai](https://itwinai.readthedocs.io/v0.3.4/tutorials/profiling/profiling-overview.html).
+
+To collect scalability metrics set the following fields in the trainer
+configuration:
+
+```yaml
+measure_gpu_data: False
[…]
