### 2025-12-05 da1dd65
Refactor trainer and localhost scalability report
