+++ joss/reproducibility/trainer.py
@@ -9,32 +9,30 @@
 # - Matteo Bunino <matteo.bunino@cern.ch> - CERN
 # --------------------------------------------------------------------------------------
 
-
 import os
-import time
-from timeit import default_timer as timer
 from typing import Any, Dict, Literal, Optional, Tuple
 
-import numpy as np
 import torch
 import torch.nn as nn
-from ray import tune
 from torch.utils.data import Dataset, TensorDataset
-from tqdm import tqdm
 
 from itwinai.distributed import suppress_workers_print
 from itwinai.loggers import Logger
 from itwinai.torch.config import TrainingConfiguration
-from itwinai.torch.distributed import DeepSpeedStrategy, RayDDPStrategy, RayDeepSpeedStrategy
-from itwinai.torch.monitoring.monitoring import measure_gpu_utilization
-from itwinai.torch.profiling.profiler import profile_torch_trainer
-from itwinai.torch.trainer import RayTorchTrainer, TorchTrainer
+from itwinai.torch.trainer import TorchTrainer
+
 from src.model import Decoder, Decoder_2d_deep, GeneratorResNet, UNet
 from src.utils import init_weights
 
 
 class VirgoTrainingConfiguration(TrainingConfiguration):
-    """Virgo TrainingConfiguration"""
+    """Virgo TrainingConfiguration (extends base TrainingConfiguration).
+
+    Additional fields:
+    - save_best: whether to save best model on validation dataset.
+    - loss: 'l1' or 'l2' for this specific use case.
+    - generator: which generator architecture to use.
+    """
 
     #: Whether to save best model on validation dataset. Defaults to True.
     save_best: bool = True
@@ -45,43 +43,71 @@ class VirgoTrainingConfiguration(TrainingConfiguration):
 
 
 class NoiseGeneratorTrainer(TorchTrainer):
+    """Trainer for Virgo noise generator models.
+
+    Reuses the generic TorchTrainer:
+    - uses its Ray / checkpoint / logging / profiling logic
+    - customizes model, loss, dataloaders and train/validation steps
+    """
+
     def __init__(
         self,
         num_epochs: int = 2,
-        config: Dict | TrainingConfiguration | None = None,
+        config: Dict | VirgoTrainingConfiguration | None = None,
         strategy: Literal["ddp", "deepspeed", "horovod"] | None = "ddp",
-        checkpoint_path: str = "checkpoints/epoch_{}.pth",
+        checkpoint_path: str = "checkpoints/epoch_{}.pth",  # kept for backwards compat
         logger: Logger | None = None,
         random_seed: int | None = None,
         name: str | None = None,
-        validation_every: int = 0,
+        validation_every: int | None = None,
+        checkpoint_every: int | None = None,
         **kwargs,
     ) -> None:
+        # Normalize config to our custom VirgoTrainingConfiguration
+        if config is None:
+            config = {}
+        if isinstance(config, dict):
+            config = VirgoTrainingConfiguration(**config)
+        elif not isinstance(config, VirgoTrainingConfiguration):
+            # If someone passes a plain TrainingConfiguration, upgrade it
+            config = VirgoTrainingConfiguration(**config.model_dump())
+
+        # Map old `checkpoint_path` to the new `checkpoints_location` directory
+        # Old default was "checkpoints/epoch_{}.pth" -> we keep only the directory.
+        if ("{" in checkpoint_path) or checkpoint_path.endswith(".pth"):
+            ckpt_root = os.path.dirname(checkpoint_path) or "checkpoints"
+        else:
+            ckpt_root = checkpoint_path
+
+        # If user passes validation_every, use it as checkpoint_every unless overridden
+        if checkpoint_every is None and validation_every is not None:
[…]
