### 2025-12-05 c6d166e
Attempt fixing linter errors
