### 2025-12-06 b9d4364
Attempt fixing linter
