+++ joss/reproducibility/trainer.py
@@ -9,20 +9,21 @@
 # - Matteo Bunino <matteo.bunino@cern.ch> - CERN
 # --------------------------------------------------------------------------------------
 
+# ruff: noqa: I001
+
 from pathlib import Path
 from typing import Any, Dict, Literal, Tuple
 
 import torch
 import torch.nn as nn
-from src.model import Decoder, Decoder2dDeep, GeneratorResNet, UNet
-from src.utils import init_weights
-from torch.utils.data import Dataset, TensorDataset
-
 from itwinai.distributed import suppress_workers_print
 from itwinai.loggers import Logger
 from itwinai.torch.config import TrainingConfiguration
 from itwinai.torch.trainer import TorchTrainer
+from torch.utils.data import Dataset, TensorDataset
 
+from src.model import Decoder, Decoder2dDeep, GeneratorResNet, UNet
+from src.utils import init_weights
 
 class VirgoTrainingConfiguration(TrainingConfiguration):
     """Virgo TrainingConfiguration (extends base TrainingConfiguration).
