### 2025-12-06 6ca87a9
Refactor Dockerfile
