### 2025-12-06 642b40b
Small refactor
