+++ joss/reproducibility/synthetic-data-gen/data_generation_hdf5.sh
@@ -1,5 +1,14 @@
 #!/bin/bash
 
+# --------------------------------------------------------------------------------------
+# Part of the interTwin Project: https://www.intertwin.eu/
+#
+# Created by: Jarl Sondre Sæther
+#
+# Credit:
+# - Jarl Sondre Sæther <jarl.sondre.saether@cern.ch> - CERN
+# --------------------------------------------------------------------------------------
+
 #SBATCH --account=intertwin
 #SBATCH --output=array-job/job_%a.out
 #SBATCH --error=array-job/job_%a.err
