+++ joss/reproducibility/hpo.py
@@ -1,226 +1,0 @@
-# --------------------------------------------------------------------------------------
-# Part of the interTwin Project: https://www.intertwin.eu/
-#
-# Created by: Anna Lappe
-#
-# Credit:
-# - Anna Lappe <anna.elisa.lappe@cern.ch> - CERN
-# - Matteo Bunino <matteo.bunino@cern.ch> - CERN
-# --------------------------------------------------------------------------------------
-
-import argparse
-from typing import Dict
-
-import matplotlib.pyplot as plt
-import ray
-import torch
-from hydra import compose, initialize
-from ray import tune
-
-from itwinai.cli import exec_pipeline_with_compose
-
-
-def run_trial(config: Dict, data: Dict):
-    """Execute a single trial using the given configuration (config).
-    This runs a full training pipeline - you can also specify a pipeline as a dictionary,
-    e.g. if you only want to run certain parts without changing your config.yaml file
-    (see below).
-
-    Args:
-        config (dict): A dictionary containing hyperparameters, such as:
-            - 'batch_size' (int): The size of the batch for training.
-            - 'lr' (float): The learning rate for the optimizer.
-        data (dict): A dictionary containing a "pipeline_name" field,
-            which specifies the training pipeline to be used.
-            Must be defined in a file called "config.yaml"
-
-    You can also run a manual pipeline by directly creating it with imported classes:
-
-    Example:
-        >>> my_pipeline = Pipeline(
-        >>>    [
-        >>>        TimeSeriesDatasetSplitter(
-        >>>            train_proportion=0.9,
-        >>>            root_folder="/p/scratch/intertwin/datasets/virgo"
-        >>>        ),
-        >>>        NoiseGeneratorTrainer(
-        >>>            config=config,
-        >>>            num_epochs=4,
-        >>>            strategy=None,
-        >>>            checkpoint_path='checkpoints/checkpoint_epoch_{}.pth',
-        >>>            validation_every=20
-        >>>        )
-        >>>    ]
-        >>> )
-
-        >>> my_pipeline.execute()
-        ```
-
-    Note: Passing a seed to TimeSeriesDatasetSplitter and NoiseGeneratorTrainer will make runs
-    uniform across trials, reducing the variablility to only the hyperparameter settings
-    """
-
-    pipe_key = data["pipeline_name"]
-
-    with initialize():
-        cfg = compose(
-            "config.yaml",
-            overrides=[
-                f"batch_size={config['batch_size']}",
-                f"optim_lr={config['optim_lr']}",
-                f"+pipe_key={pipe_key}",
-            ],
-        )
-        exec_pipeline_with_compose(cfg)
-
-
-def run_hpo(args):
-    """Run hyperparameter optimization using Ray Tune.
-    Either starts a new optimization run or resumes from previous results.
-
-    Args:
-    - args: Command-line arguments parsed by argparse.
-    """
-    if not args.load_old_results:
-        # Initialize Ray with cluster configuration from environment variables
-        ray.init(address="auto")
-
-        # Define the search space for hyperparameters
-        search_space = {
-            "batch_size": tune.choice([3, 4, 5, 6]),
-            "lr": tune.uniform(1e-5, 1e-3),
-        }
-
-        tune_config = tune.TuneConfig(
-            metric=args.metric,  # Metric to optimize (loss by default)
-            mode="min",  # Minimize the loss
-            num_samples=args.num_samples,  # Number of trials to run
-        )
-
-        run_config = tune.RunConfig(
-            name="Virgo-Ray-Experiment", stop={"training_iteration": args.max_iterations}
-        )
-
-        # Determine GPU and CPU utilization per trial
[…]
