+++ joss/reproducibility/synthetic-data-gen/data_generation_small.sh
@@ -1,5 +1,14 @@
 #!/bin/bash
 
+# --------------------------------------------------------------------------------------
+# Part of the interTwin Project: https://www.intertwin.eu/
+#
+# Created by: Matteo Bunino
+#
+# Credit:
+# - Matteo Bunino <matteo.bunino@cern.ch> - CERN
+# --------------------------------------------------------------------------------------
+
 #SBATCH --account=intertwin
 #SBATCH --output=array-job/job_%a.out
 #SBATCH --error=array-job/job_%a.err
