### 2025-12-06 d9ee46f
Draft HPC scripts
