+++ joss/reproducibility/scalability/slurm.jsc.sh
@@ -44,7 +44,8 @@ echo "DEBUG: SLURM_CPUS_PER_TASK: $SLURM_CPUS_PER_TASK"
 echo "DEBUG: TRAINING_CMD: $TRAINING_CMD"
 
 # Setup env for distributed ML
-export CUDA_VISIBLE_DEVICES=$(seq -s, 0 $((SLURM_GPUS_PER_NODE - 1)))
+CUDA_VISIBLE_DEVICES=$(seq -s, 0 $((SLURM_GPUS_PER_NODE - 1)))
+export CUDA_VISIBLE_DEVICES
 echo "DEBUG: CUDA_VISIBLE_DEVICES: $CUDA_VISIBLE_DEVICES"
 export OMP_NUM_THREADS=1
 if [ "$SLURM_GPUS_PER_NODE" -gt 0 ] ; then
