### 2025-12-08 aa4dca5
Joss reproducibility (#466)
* Add new references for scalability and profiling

* Add reproducibility section to paper

Added a section on reproducibility detailing the scalability results obtained on the JUWELS Booster GPU partition, including configuration file management and reproducibility of results.

* Update URLs in bibliography for versioning

* Draft Virgo code

* Code cleanup

* Add profiling metrics to pipeline

* First version datagen tutorial

* Fix bind mount

* Refactor trainer and localhost scalability report

* Attempt fixing linter errors

* Attempt fixing linter

* Ignore im
[…]
