+++ joss/paper.bib
@@ -233,3 +233,23 @@ This data was used to generate the plots in the JOSS submission for itwinai},
   year={2019},
   doi     = {10.17815/jlsrf-5-171},
 }
+
+@misc{itwinai_docs_scalability,
+  author    = {itwinai Developers},
+  title     = {Scalability report},
+  year      = {2025},
+  publisher = {Read the Docs},
+  journal   = {itwinai documentation},
+  url       = {https://itwinai.readthedocs.io/v0.3.4/how-it-works/scalability-report/scalability_report.html},
+  urldate   = {2025-08-14},
+}
+
+@misc{itwinai_docs_profiling,
+  author    = {itwinai Developers},
+  title     = {Profiling overview},
+  year      = {2025},
+  publisher = {Read the Docs},
+  journal   = {itwinai documentation},
+  url       = {https://itwinai.readthedocs.io/v0.3.4/tutorials/profiling/profiling-overview.html},
+  urldate   = {2025-08-14},
+}
