### 2025-12-09 interTwin-eu/itwinai#467 @matbun
Profiling errors in CPU-only workflows
Please refer to this: https://github.com/openjournals/joss-reviews/issues/9409#issuecomment-3629476192

To reproduce, the code (single-host env) is: https://github.com/interTwin-eu/itwinai/tree/joss/joss/reproducibility
