### 2025-12-09 #9409 @matbun
Thanks @yewentao256  for the feedback!
Being designed for large-scale AI workloads, we developed and tested our software mostly on GPU platforms, so there may be some problems when running on CPU-only nodes. This is definitely something on which we could improve.
When writing the reproduciblity guide I also bumped into the second error... it seems to be related to race conditions and occurs only some times. If you want you can try re-running the code and it should work.

Not to forget about this, I have reported this here: https://github.com/interTwin-eu/itwinai/issues/467

Hopefully this will not be a problem for this review :)
