### 2025-12-09 interTwin-eu/itwinai#468 @xiazeyu
add a generic section in the installation document
In these two pages:
https://itwinai.readthedocs.io/latest/installation/user_installation.html
https://itwinai.readthedocs.io/latest/installation/developer_installation.html

1. would you consider to add the requirements (or required modules) for generic HPC? So we know which packages/versions are necessary and must be load from ml; which is optional and might be able to load later via pip.

In our university, we use [Rivanna](https://www.rc.virginia.edu/userinfo/hpc/). We have a very different module list than these provided clusters. A list of available modules can be found [here](https://www.rc.virginia.edu/userinfo/hpc/software/complete-list/).

For Juelich (JSC), PyTorch:
```shell
ml --force purge
ml Stages/2025 # what is this for? Rivanna don't have this package.
ml GCC OpenMPI git # any specific requirements to package version? in Rivanna, openmpi is included in gompi.
ml CUDA/12 # I guess the cuda version depends on PyTorch/Tensorflow, but it would be good to state if it has to be 12, if there's sub-version requirement, or if it works for cuda 11 as well.
ml cuDNN
ml MPI-settings/CUDA # what's the difference of this cuda to the cuda above?
ml Python CMake HDF5 PnetCDF libaio mpi4py
ml libaio # did not find this in rivanna
```

For Juelich (JSC), TensorFlow:
I saw Stages have different version (2024), and many dependencies have detailed version requirement.

Vega:
This looks close to the packages provided by Rivanna, but are these version strict? I also noticed additonal dependencies of nccl.

LUMI:
Could you provide some brief explaination why Python virtual environments are discouraged in favour of containers on this cluster?

My command for installation for your reference:
```shell
ml --force purge
ml gcc gompi cudnn python cmake hdf5 pnetcdf mpi4py git uv nccl cuda
```

2. Since dependencies are the same for PyTorch and TensorFlow in Vega, I would suggest to leave one tab.

ref https://github.com/openjournals/joss-reviews/issues/9409
