### 2025-12-10 interTwin-eu/itwinai#468 @xiazeyu
4. Will it be better to refer to a stable version for Horovod and Microsoft DeepSpeed, instead of their main branch?

I encountered the second issue mentioned in https://github.com/horovod/horovod/pull/4297.
Installation of a pre-built wheel typically bypass problems like this one. But let me know if you think building on main branch is better for the user.
