### 2025-12-11 interTwin-eu/itwinai#470 @xiazeyu
bump version to solve "No such command 'run'."
I was trying to reproduce the MNIST example shown in https://itwinai.readthedocs.io/latest/getting-started/complete-workflow-example.html#mnist-example.

And I received the error:
```
(itwinai-torch) bash-4.4$itwinai run -c https://raw.githubusercontent.com/interTwin-eu/itwinai/refs/heads/main/use-cases/mnist/torch/run-example.yaml
Usage: itwinai [OPTIONS] COMMAND [ARGS]...
Try 'itwinai --help' for help.
╭─ Error ────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────╮
│ No such command 'run'.                                                                                                                                                                                                                                                                     │
╰────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────╯
```

I think it is due to the latest release https://github.com/interTwin-eu/itwinai/releases/tag/v0.3.4 did not have the run command, which was added to the main branch in https://github.com/interTwin-eu/itwinai/pull/456.

I would suggest to bump the version to eliminate ambiguity.

Additionally, would you add some document to explain the difference between `itwinai exec-pipeline --config-name config.yaml` using [config.yaml](https://github.com/interTwin-eu/itwinai/blob/joss/use-cases/mnist/torch/config.yaml) and `itwinai run -c run_config.yaml` using [run_config.yaml](https://github.com/interTwin-eu/itwinai/blob/joss/use-cases/mnist/torch/run-example.yaml)?

ref https://github.com/openjournals/joss-reviews/issues/9409
