### 2025-12-11 interTwin-eu/itwinai#471 @xiazeyu
complete-workflow-example did not submit job
I used the commandline provided [here](https://itwinai.readthedocs.io/latest/getting-started/complete-workflow-example.html#mnist-example), and I found itwinai generated the slurm script, but did not submit it.

In the section [Expected Output](https://itwinai.readthedocs.io/latest/getting-started/complete-workflow-example.html#expected-output), step 3 says it will "Submit the job to your HPC cluster", but the command will not do so unless explicitly provide `-j` option.

Please let me know if this is expected behaviour or not. I did not see that is documented, and in `run_example.yaml`, there is an option of `slurm_config.submit_job` that is likely to control the submission behaviour.

ref https://github.com/openjournals/joss-reviews/issues/9409
