### 2025-12-11 interTwin-eu/itwinai#472 @xiazeyu
unable to execute on UVA Rivanna
I followed the tutorial in https://itwinai.readthedocs.io/latest/getting-started/complete-workflow-example.html, and did some modification trying to make it runs on UVA Rivanna cluster.

changes made:
https://github.com/interTwin-eu/itwinai/blob/main/use-cases/mnist/torch/run-example.yaml

```yaml
slurm_config:
  account: <my own account>
  partition: <my own partition>
  pre_exec_file: src/itwinai/slurm/system-base-scripts/rivanna_pre_exec.sh

```

new file `src/itwinai/slurm/system-base-scripts/rivanna_pre_exec.sh`:
copied from [`juwels_pre_exec.sh`](https://github.com/interTwin-eu/itwinai/blob/main/src/itwinai/slurm/system-base-scripts/juwels_pre_exec.sh)

```shell
# Load environment modules
ml --force purge
ml gcc gompi cudnn python cmake hdf5 pnetcdf mpi4py git uv nccl cuda

# removed previous loading commands
```

Output of `itwinai run -c use-cases/mnist/torch/run-example.yaml`

<summary>Log</summary>

```
Installing 'git+https://github.com/matbun/itwinai-mnist-plugin.git'
#################### SLURM Script Preview ####################
#!/bin/bash

# Job configuration
#SBATCH --job-name=ddp-2x4
#SBATCH --account=bi_dsc_community
#SBATCH --partition=bii-gpu
#SBATCH --time=00:30:00

#SBATCH --output=slurm-job-logs/ddp-2x4.out
#SBATCH --error=slurm-job-logs/ddp-2x4.err

# Resource allocation
#SBATCH --nodes=2
#SBATCH --ntasks-per-node=1
#SBATCH --cpus-per-task=16
#SBATCH --gpus-per-node=4
#SBATCH --gres=gpu:4
#SBATCH --mem=64G

# Load environment modules
ml --force purge
ml gcc gompi cudnn python cmake hdf5 pnetcdf mpi4py git uv nccl cuda

# Job info
echo "DEBUG: TIME: $(date)"
echo "Running on system: $SYSTEMNAME"
echo "DEBUG: EXECUTE: $EXEC"
echo "DEBUG: SLURM_SUBMIT_DIR: $SLURM_SUBMIT_DIR"
echo "DEBUG: SLURM_JOB_ID: $SLURM_JOB_ID"
echo "DEBUG: SLURM_JOB_NODELIST: $SLURM_JOB_NODELIST"
echo "DEBUG: SLURM_NNODES: $SLURM_NNODES"
echo "DEBUG: SLURM_NTASKS: $SLURM_NTASKS"
echo "DEBUG: SLURM_TASKS_PER_NODE: $SLURM_TASKS_PER_NODE"
echo "DEBUG: SLURM_SUBMIT_HOST: $SLURM_SUBMIT_HOST"
echo "DEBUG: SLURMD_NODENAME: $SLURMD_NODENAME"
echo "DEBUG: SLURM_CPUS_PER_TASK: $SLURM_CPUS_PER_TASK"
echo "DEBUG: TRAINING_CMD: $TRAINING_CMD"

if [ "$DEBUG" = true ] ; then
  echo "DEBUG: NCCL_DEBUG=INFO"
  export NCCL_DEBUG=INFO
fi
echo

# Setup env for distributed ML
CUDA_VISIBLE_DEVICES=$(seq -s, 0 $((SLURM_GPUS_PER_NODE - 1)))
export CUDA_VISIBLE_DEVICES
echo "DEBUG: CUDA_VISIBLE_DEVICES: $CUDA_VISIBLE_DEVICES"

OMP_NUM_THREADS=1
if [ "$SLURM_GPUS_PER_NODE" -gt 0 ] ; then
  OMP_NUM_THREADS=$((SLURM_CPUS_PER_TASK / SLURM_GPUS_PER_NODE))
fi
export OMP_NUM_THREADS

# Prevent NCCL not figuring out how to initialize.
export NCCL_SOCKET_IFNAME=ib0
# Prevent Gloo not being able to communicate.
export GLOO_SOCKET_IFNAME=ib0

function torchrun-launcher(){
  MASTER_ADDR="$(scontrol show hostnames "$SLURM_JOB_NODELIST" | head -n 1)i"
  MASTER_PORT=54123
  export MASTER_ADDR MASTER_PORT
  torchrun_command='torchrun_jsc \
    --log_dir="logs_torchrun" \
    --nnodes="$SLURM_NNODES" \
    --nproc_per_node="$SLURM_GPUS_PER_NODE" \
    --rdzv_id="$SLURM_JOB_ID" \
    --rdzv_conf=is_host="$(( SLURM_NODEID == 0 ? 1 : 0 ))" \
    --rdzv_backend=c10d \
    --rdzv_endpoint="$MASTER_ADDR":"$MASTER_PORT" \
    --no-python'
  torchrun_command="$torchrun_command $1"
  srun --cpu-bind=none --ntasks-per-node=1 bash -c "$torchrun_command"
}

function torchrun-launcher-container(){
  MASTER_ADDR="$(scontrol show hostnames "$SLURM_JOB_NODELIST" | head -n 1)i"
  MASTER_PORT=54123
  export MASTER_ADDR MASTER_PORT
  torchrun_command='torchrun_jsc \
    --log_dir="logs_torchrun" \
    --nnodes="$SLURM_NNODES" \
    --nproc_per_node="$SLURM_GPUS_PER_NODE" \
    --rdzv_id="$SLURM_JOB_ID" \
    --rdzv_conf=is_host="$(( SLURM_NODEID == 0 ? 1 : 0 ))" \
    --rdzv_backend=c10d \
    --rdzv_endpoint="$MASTER_ADDR":"$MASTER_PORT" \
    --no-python'
  torchrun_command="$torchrun_command $1"
  srun --cpu-bind=none --ntasks-per-node=1 "$CONTAINER_PATH" /bin/bash -c "$torchrun_command"
}

function py-spy-torchrun-launcher(){
  local rate="$1"
  local output="$2"

  MASTER_ADDR="$(scontrol show hostnames "$SLURM_JOB_NODELIST" | head -n 1)i"
  MASTER_PORT=54123
  export MASTER_ADDR MASTER_PORT

  srun --cpu-bind=none --ntasks-per-node=1 \
    bash -c 'py-spy record -r "$0" -s -o "$1" -f raw -- torchrun_jsc \
    --log_dir="logs_torchrun" \
    --nnodes="$SLURM_NNODES" \
    --nproc_per_node="$SLURM_GPUS_PER_NODE" \
    --rdzv_id="$SLURM_JOB_ID" \
    --rdzv_conf=is_host="$(( SLURM_NODEID == 0 ? 1 : 0 ))" \
    --rdzv_backend=c10d \
    --rdzv_endpoint="$MASTER_ADDR":"$MASTER_PORT" \
    --no-python \
    "$@"' "$rate" "$output" "$3"
}

function py-spy-torchrun-launcher-container(){
  local rate="$1"
  local output="$2"

  MASTER_ADDR="$(scontrol show hostnames "$SLURM_JOB_NODELIST" | head -n 1)i"
  MASTER_PORT=54123
  export MASTER_ADDR MASTER_PORT

  srun --cpu-bind=none --ntasks-per-node=1 \
    "$CONTAINER_PATH" /bin/bash -c 'py-spy record -r "$0" -s -o "$1" -f raw -- torchrun_jsc \
    --log_dir="logs_torchrun" \
    --nnodes="$SLURM_NNODES" \
    --nproc_per_node="$SLURM_GPUS_PER_NODE" \
    --rdzv_id="$SLURM_JOB_ID" \
    --rdzv_conf=is_host="$(( SLURM_NODEID == 0 ? 1 : 0 ))" \
    --rdzv_backend=c10d \
    --rdzv_endpoint="$MASTER_ADDR":"$MASTER_PORT" \
    --no-python \
    "$@"' "$rate" "$output" "$3"
}

function srun-launcher(){
  srun --cpu-bind=none --ntasks-per-node="$SLURM_GPUS_PER_NODE" \
    --cpus-per-task=$((SLURM_CPUS_PER_TASK / SLURM_GPUS_PER_NODE)) \
    --ntasks=$((SLURM_GPUS_PER_NODE * SLURM_NNODES)) \
    bash -c "$1"
}

# Get GPUs info per node
srun --cpu-bind=none --ntasks-per-node=1 \
  bash -c 'echo -e "NODE hostname: $(hostname)\n$(nvidia-smi)\n\n"'

export ITWINAI_LOG_LEVEL="DEBUG"

function srun-launcher-container(){
  srun --cpu-bind=none --ntasks-per-node="$SLURM_GPUS_PER_NODE" \
    --cpus-per-task=$((SLURM_CPUS_PER_TASK / SLURM_GPUS_PER_NODE)) \
    --ntasks=$((SLURM_GPUS_PER_NODE * SLURM_NNODES)) \
      "$CONTAINER_PATH" /bin/bash -c "$1"
}

# Get GPUs info per node
srun --cpu-bind=none --ntasks-per-node=1 \
  bash -c 'echo -e "NODE hostname: $(hostname)\n$(nvidia-smi)\n\n"'

export ITWINAI_LOG_LEVEL="DEBUG"

function ray-launcher(){
  num_gpus=$SLURM_GPUS_PER_NODE
  num_cpus=$SLURM_CPUS_PER_TASK

  # This tells Tune to not change the working directory to the trial directory
  # which makes relative paths accessible from inside a trial
  export RAY_CHDIR_TO_TRIAL_DIR=0
  export RAY_DEDUP_LOGS=0
  export RAY_USAGE_STATS_DISABLE=1

  # Disable colors in output
  export NO_COLOR=1
  export RAY_COLOR_PREFIX=0

  #########   Set up Ray cluster   ########

  # Get the node names
  nodes=$(scontrol show hostnames "$SLURM_JOB_NODELIST")
  mapfile -t nodes_array <<< "$nodes"

  # The head node will act as the central manager (head) of the Ray cluster.
  head_node=${nodes_array[0]}
  port=7639       # This port will be used by Ray to communicate with worker nodes.

  echo "Starting HEAD at $head_node"
  # Start Ray on the head node.
  # The `--head` option specifies that this node will be the head of the Ray cluster.
  # `srun` submits a job that runs on the head node to start the Ray head with the specified
  # number of CPUs and GPUs.
  srun --nodes=1 --ntasks=1 -w "$head_node" \
      ray start --head --node-ip-address="$head_node"i --port=$port \
      --num-cpus "$num_cpus" --num-gpus "$num_gpus"  --block &

  # Wait for a few seconds to ensure that the head node has fully initialized.
  sleep 1

  echo HEAD node started.

  # Start Ray worker nodes
  # These nodes will connect to the head node and become part of the Ray cluster.
  worker_num=$((SLURM_JOB_NUM_NODES - 1))    # Total number of worker nodes (excl the head node)
  for ((i = 1; i <= worker_num; i++)); do
      node_i=${nodes_array[$i]}   # Get the current worker node hostname.
      echo "Starting WORKER $i at $node_i"

      # Use srun to start Ray on the worker node and connect it to the head node.
      # The `--address` option tells the worker node where to find the head node.
      srun --nodes=1 --ntasks=1 -w "$node_i" \
          ray start --address "$head_node"i:"$port" --redis-password='5241580000000000' \
          --num-cpus "$num_cpus" --num-gpus "$num_gpus" --block &

      sleep 5 # Wait before starting the next worker to prevent race conditions.
  done
  echo All Ray workers started.

  # Run command without srun
  # if you want the number of workers to be adaptive during distributed training append this:
  # training_pipeline.steps.training_step.ray_scaling_config.num_workers=$(($SLURM_GPUS_PER_NODE * $SLURM_NNODES))
  $1
}

function ray-launcher-container(){
  num_gpus=$SLURM_GPUS_PER_NODE
  num_cpus=$SLURM_CPUS_PER_TASK

  # This tells Tune to not change the working directory to the trial directory
  # which makes relative paths accessible from inside a trial
  export RAY_CHDIR_TO_TRIAL_DIR=0
  export RAY_DEDUP_LOGS=0
  export RAY_USAGE_STATS_DISABLE=1

  # Disable colors in output
  export NO_COLOR=1
  export RAY_COLOR_PREFIX=0

  #########   Set up Ray cluster   ########

  # Get the node names
  nodes=$(scontrol show hostnames "$SLURM_JOB_NODELIST")
  mapfile -t nodes_array <<< "$nodes"

  # The head node will act as the central manager (head) of the Ray cluster.
  head_node=${nodes_array[0]}
  port=7639       # This port will be used by Ray to communicate with worker nodes.

  echo "Starting HEAD at $head_node"
  # Start Ray on the head node.
  # The `--head` option specifies that this node will be the head of the Ray cluster.
  # `srun` submits a job that runs on the head node to start the Ray head with the specified
  # number of CPUs and GPUs.
  srun --nodes=1 --ntasks=1 -w "$head_node" \
      ray start --head --node-ip-address="$head_node"i --port=$port \
      --num-cpus "$num_cpus" --num-gpus "$num_gpus"  --block &

  # Wait for a few seconds to ensure that the head node has fully initialized.
  sleep 1

  echo HEAD node started.

  # Start Ray worker nodes
  # These nodes will connect to the head node and become part of the Ray cluster.
  worker_num=$((SLURM_JOB_NUM_NODES - 1))    # Total number of worker nodes (excl the head node)
  for ((i = 1; i <= worker_num; i++)); do
      node_i=${nodes_array[$i]}   # Get the current worker node hostname.
      echo "Starting WORKER $i at $node_i"

      # Use srun to start Ray on the worker node and connect it to the head node.
      # The `--address` option tells the worker node where to find the head node.
      srun --nodes=1 --ntasks=1 -w "$node_i" \
          ray start --address "$head_node"i:"$port" --redis-password='5241580000000000' \
          --num-cpus "$num_cpus" --num-gpus "$num_gpus" --block &

      sleep 5 # Wait before starting the next worker to prevent race conditions.
  done
  echo All Ray workers started.

  # Run command without srun
  # if you want the number of workers to be adaptive during distributed training append this:
  # training_pipeline.steps.training_step.ray_scaling_config.num_workers=$(($SLURM_GPUS_PER_NODE * $SLURM_NNODES))
  "$CONTAINER_PATH" /bin/bash -c "$1"
}

torchrun-launcher 'itwinai exec-pipeline \
--config-name=config \
--config-path=. \
+pipe_key=rnn_training_pipeline \
strategy=ddp \
experiment_name=main-experiment \
run_name=main-run'
##############################################################

```

After executing `itwinai run -c use-cases/mnist/torch/run-example.yaml -j`:

```
bash-4.4$itwinai run -c use-cases/mnist/torch/run-example.yaml -j
Installing 'git+https://github.com/matbun/itwinai-mnist-plugin.git'
Submitted batch job 6666902
```

ddp-2x4.out:

```
DEBUG: TIME: Thu Dec 11 02:29:02 EST 2025
Running on system:
DEBUG: EXECUTE:
DEBUG: SLURM_SUBMIT_DIR: /sfs/gpfs/tardis/home/yxn7cj/itwinai-dev
DEBUG: SLURM_JOB_ID: 6666903
DEBUG: SLURM_JOB_NODELIST: udc-aj36-35,udc-aj40-35
DEBUG: SLURM_NNODES: 2
DEBUG: SLURM_NTASKS: 2
DEBUG: SLURM_TASKS_PER_NODE: 1(x2)
DEBUG: SLURM_SUBMIT_HOST: udc-aj38-35
DEBUG: SLURMD_NODENAME: udc-aj36-35
DEBUG: SLURM_CPUS_PER_TASK: 16
DEBUG: TRAINING_CMD:

DEBUG: CUDA_VISIBLE_DEVICES: 0,1,2,3

```

ddp-2x4.err:

```
srun: error: udc-aj36-35: task 0: Exited with exit code 2
srun: Terminating StepId=6666903.0
[2025-12-11T02:29:03.160] error: execve(): bash: No such file or directory
[2025-12-11T02:29:03.321] error: execve(): bash: No such file or directory
srun: error: udc-aj40-35: task 1: Exited with exit code 2
srun: error: udc-aj36-35: task 0: Exited with exit code 2
srun: Terminating StepId=6666903.1
[2025-12-11T02:29:03.477] error: execve(): bash: No such file or directory
[2025-12-11T02:29:03.520] error: execve(): bash: No such file or directory
srun: error: udc-aj40-35: task 1: Exited with exit code 2
[2025-12-11T02:29:03.736] error: execve(): bash: No such file or directory
srun: error: udc-aj36-35: task 0: Exited with exit code 2
srun: Terminating StepId=6666903.2
srun: error: udc-aj40-35: task 1: Exited with exit code 2
[2025-12-11T02:29:03.756] error: execve(): bash: No such file or directory

```

I did some sanity check by `itwinai sanity-check`:
```
(itwinai-dev) bash-4.4$itwinai sanity-check
✅ Successfully imported: itwinai
✅ Successfully imported: itwinai.cli
✅ Successfully imported: itwinai.components
✅ Successfully imported: itwinai.distributed
✅ Successfully imported: itwinai.loggers
✅ Successfully imported: itwinai.parser
✅ Successfully imported: itwinai.pipeline
✅ Successfully imported: itwinai.serialization
✅ Successfully imported: itwinai.type
✅ Successfully imported: itwinai.utils
✅ Successfully imported: itwinai.tests
✅ Successfully imported: itwinai.tests.dummy_components
✅ Successfully imported: itwinai.tests.exceptions
✅ Successfully imported: itwinai.tests.sanity_check
✅ Successfully imported: itwinai.plugins
✅ Successfully imported: itwinai.scalability_report
✅ Successfully imported: itwinai.scalability_report.data
✅ Successfully imported: itwinai.scalability_report.utils
✅ Successfully imported: itwinai.scalability_report.reports
✅ Successfully imported: itwinai.scalability_report.plot
✅ Successfully imported: itwinai.constants

All modules imported successfully!
```

ref https://github.com/openjournals/joss-reviews/issues/9409
