### 2025-12-11 interTwin-eu/itwinai#473 @xiazeyu
joss reference
I would recommend to replace the link for "Accelerating the Machine Learning Lifecycle with MLflow." with http://sites.computer.org/debull/A18dec/p39.pdf

And replace "Scalability analysis of GlitchFlow with itwinai" with its doi: https://doi.org/10.5281/zenodo.16882390

ref https://github.com/openjournals/joss-reviews/issues/9409
