### 2025-12-11 4100d88
Update installation commands for GPU acceleration
