### 2025-12-11 d753b1c
Update URL to DOI in paper.bib entry
