### 2025-12-11 interTwin-eu/itwinai#472 @xiazeyu
Thanks for that. I will try `ml bash` to see if that works. But please let me know if you would like me to try this function, or if there's some other examples that can be tested more easily.

I've checked the test cases and that seems to be a good proof for itwinai to be usable. But a lot of the code are written for these three HPC clusters, and I am not able to test on my own cluster. (I don't expect any barrier that would prevent it to do so). The slurm submission is working and the slurm script seems to generate in a correct manner. I would assume it's just the environment is configured in a different way.

I thought it will be much powerful if itwinai can be used on other clusters though.
