+++ joss/paper.bib
@@ -176,7 +176,7 @@ This data was used to generate the plots in the JOSS submission for itwinai},
 	journal = {IEEE Data Eng. Bull.},
 	pages = {39--45},
 	title = {{Accelerating the Machine Learning Lifecycle with MLflow}},
-	url = {https://api.semanticscholar.org/CorpusID:83459546},
+	url = {http://sites.computer.org/debull/A18dec/p39.pdf},
 	volume = {41},
 	year = {2018}
 }
