### 2025-12-11 interTwin-eu/itwinai#470 @matbun
> I would suggest to make a stable branch that is always synchrnoized with its stable (or numbered) version of the readthedocs and the latest released package.
>
> If you decided not to include documentation for `itwinai run`, that is fine. But I would expect the default version for readthedocs to be aligned with the latest released package.

The docs are already in sync with the releases. The `itwinai run` command is not part of any "versioned" doc page because it is not part of any release either. It is only on the main branch for the moment. It will be included in the next software release and in the related versioned documentation. Maybe this is the "stable" docs page you refer to? https://itwinai.readthedocs.io/stable/

Maybe the source of the confusion comes from the fact that the default docs page is the "latest" (pointing to main), while when you install from PyPI you get the latest itwinai version (the "stable"). Changing the default docs page from "[latest](https://itwinai.readthedocs.io/latest/)" to "[stable](https://itwinai.readthedocs.io/stable/)" would solve this misunderstanding?

> For the difference, I would suggest to discuss it here https://github.com/interTwin-eu/itwinai/blob/joss/use-cases/mnist/torch/README.md, and maybe here as well https://itwinai.readthedocs.io/latest/getting-started/complete-workflow-example.html#mnist-example.
>
> The discussion can be short, and you can refer to their API references as well (honestly API reference will not be the first place if I'm trying in the "Quick Start" chapter). The main point is to let the user know if `run` is a upgraded version for `exec-pipeline`, not a replacement. And it is also in development, so the latest release do not have this feature.

Okay, I'll propose some changes to better document that. But still my main question: this feature is not part of itwinai `0.3.4`, which is the target for the joss publication. Should I consider this issue as part of the joss review or just a generic (useful) feedback?
