+++ .github/workflows/container-ci.yml
@@ -58,7 +58,7 @@ jobs:
 
     steps:
       - name: Checkout Code
-        uses: actions/checkout@v5
+        uses: actions/checkout@v6
 
       - name: Build and Test with Custom Action
         uses: ./.github/actions/build-container
