### 2025-12-11 interTwin-eu/itwinai#472 @matbun
If you want to do a quick sanity check, you can `sbatch` the following SLURM job script:

```bash
#!/bin/bash

# Here add all the #SBATCH directives...

# Load environment modules
ml --force purge
ml gcc gompi cudnn python cmake hdf5 pnetcdf mpi4py git uv nccl cuda

# Load itwinai virtual env (may need to be adapted to your case)
source .venv/bin/activate

srun itwinai --help

srun torchrun --help

srun bash -c "torchrun --help"
```
