### 2025-12-12 3daab6c
Add '-j' option for SLURM job submission in examples
Updated command examples to include the '-j' option for SLURM job submission.
