### 2025-12-12 971a979
Update README with exec-pipeline command usage
Added details on executing the ML workflow using the exec-pipeline command.
