+++ use-cases/mnist/torch/README.md
@@ -9,6 +9,9 @@ use cases based on the popular MNIST dataset.
 
 It is possible to launch the training of a CNN classifier on the MNIST dataset using the
 YAML configuration file describing the whole training workflow.
+In this case, the `itwinai exec-pipeline` command is used to execute a ML workflow defined
+in the `config.yaml` file. You can find more details on this command in the [`exec-pipeline`](
+https://itwinai.readthedocs.io/latest/api/cli_reference.html#exec-pipeline) CLI reference.
 
 ```bash
 # Run the whole training pipeline
