### 2025-12-12 458d3be
Update README with training pipeline execution command
Added instructions for running the training pipeline.
