### 2025-12-12 3c8e855
Explain difference between run and exec-pipeline
