### 2025-12-12 interTwin-eu/itwinai#478 @xiazeyu
Community guidelines
> Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

Can you add these to your README and [this section](https://itwinai.readthedocs.io/stable/#community-support) of your documentation?

It can be simply like "create a [new issue](https://github.com/interTwin-eu/itwinai/issues) in GitHub if you found any issues or problems, or want to seek support". "Fork and create a [Pull Request](https://github.com/interTwin-eu/itwinai/pulls) if you want to make contribution".

Thank would be more clear for new users.

ref https://github.com/openjournals/joss-reviews/issues/9409
