### 2025-12-12 interTwin-eu/itwinai#468 @xiazeyu
> SW dependencies is one of the most complicated aspects of the deployment of itwinai. In the last three years we learned that there is no silver bullet for that and is mostly a trial and error process, and when you deploy on a new system it is common to face problems like the ones you encountered... Unfortunately, to the best of my knowledge, there is not an easy way to manage dependencies and declare version ranges as you would do for python packages in a `pyproject.toml` file.

I would highly agree with that. I would recommend to add a section in the document advising something like "while itwinai do not have requirements for certain version of CUDA, PyTorch, TensorFlow or other packages, it is common to face problems while deploying on a new system. We would encourage the user do some trial with difference versions of these software. We will (or will not, or depends on availability, will) support for these systems. Users can create a new issue or pull request to report their experience on their own systems."

I think this will make itwinai more suitable for its target audience (stated in the Statement of Need), rather than just to people who have access to these three clusters.

> We can _try_ to come up with a list of SW dependencies, e.g., CUDA, but this is probably not going to work anywhere...

In case there is no strict restriction for version for installing itwinai, you do not need to provide that. Versions can be provided only for the reproduction of examples (in case something breaks the example in the future).

But please do come up with a more generic SW dependencies.

> Also, consider that itwinai still has a low [Technology Readiness Level](https://en.wikipedia.org/wiki/Technology_readiness_level) (TRL), so the deployment is expected not to be always smooth. Hopefully, itwinai will be made more robust in future releases.

Understood. I think the project is quite complete now, but if I would expect a certain degree of generality for a submission.
