### 2025-12-12 578bee9
Add contribution section to README
Added contribution guidelines and support links.
