### 2025-12-12 23ddc99
Revise contribution guidelines in README.md
