### 2025-12-12 c498b87
Revise contribution guidelines in README.md
