### 2025-12-12 4f4074d
Add contribution section to documentation
Added contribution guidelines for the itwinai project.
