+++ README.md
@@ -38,5 +38,6 @@ at the following documents:
 Want to help improve **itwinai**? Here are a few good ways to get involved:
 
 - **Report a bug / request a feature:** open a [GitHub issue](https://github.com/interTwin-eu/itwinai/issues).
-- **Contribute code or docs:** fork the repo and submit a [pull request](https://github.com/interTwin-eu/itwinai/pulls).
-- **Ask questions or float ideas:** start a [GitHub discussion](https://github.com/interTwin-eu/itwinai/discussions) or join us on [Discord](https://discord.gg/7ru8u2nP).
+- **Contribute code or docs:** fork the repository and submit a [pull request](https://github.com/interTwin-eu/itwinai/pulls).
+- **Ask questions or float ideas:** start a [GitHub discussion](https://github.com/interTwin-eu/itwinai/discussions)
+or join us on [Discord](https://discord.gg/7ru8u2nP).
