+++ README.md
@@ -40,4 +40,4 @@ Want to help improve **itwinai**? Here are a few good ways to get involved:
 - **Report a bug / request a feature:** open a [GitHub issue](https://github.com/interTwin-eu/itwinai/issues).
 - **Contribute code or docs:** fork the repository and submit a [pull request](https://github.com/interTwin-eu/itwinai/pulls).
 - **Ask questions or float ideas:** start a [GitHub discussion](https://github.com/interTwin-eu/itwinai/discussions)
-or join us on [Discord](https://discord.gg/7ru8u2nP).
+or join us on [Discord](https://discord.gg/EZRtuK9Vjr).
