### 2025-12-12 interTwin-eu/itwinai#468 @matbun
> Understood. I think the project is quite complete now, but if I would expect a certain degree of generality for a submission.

In this case, "generality" comes from the ability of deploying itwinai to different systems. This is generally possible, it just takes some trial and error because each HPC system is different, not because of intrinsic limitations of itwinai (which is just a python library). In other words, you would probably face the same problems when using pytorch, deepspeed, horovod...

Hopefully, future work on itwinai will help reducing these issues, but there is no silver bullet unfortunately.

I'll let @r-sarma comment on this as well
