### 2025-12-15 20bc4de
Add community guidelines (#479)
* Add contribution section to README

Added contribution guidelines and support links.

* Revise contribution guidelines in README.md

* Revise contribution guidelines in README.md

* Add contribution section to documentation

Added contribution guidelines for the itwinai project.

* Fix linter checks

* Update Discord invite link
