### 2025-12-15 bb45a33
Refactor "complete workflow example" tutorial (#477)
* Add '-j' option for SLURM job submission in examples

Updated command examples to include the '-j' option for SLURM job submission.

* Update README with exec-pipeline command usage

Added details on executing the ML workflow using the exec-pipeline command.

* Update README with training pipeline execution command

Added instructions for running the training pipeline.

* Explain difference between run and exec-pipeline

* Fix formatting errors

* Add missing dot

* Fix header hierarchy

* Point to stable docs page

* Fix typo
