### 2025-12-15 607cef4
Update joss references (#474)
* Update URL to DOI in paper.bib entry

* Update URL for MLflow paper reference
