+++ joss/paper.bib
@@ -121,7 +121,7 @@
 
 @misc{saether_scalabiliy_2025,
 	title = {Scalability analysis of {GlitchFlow} with itwinai},
-	url = {https://zenodo.org/records/16882390},
+	url = {https://doi.org/10.5281/zenodo.16882390},
 	abstract = {Scalability analysis on JUWELS Booster of GlitchFlow use case (INFN, Virgo) from interTwin, using the itwinai library.
 
 This data was used to generate the plots in the JOSS submission for itwinai},
@@ -176,7 +176,7 @@ This data was used to generate the plots in the JOSS submission for itwinai},
 	journal = {IEEE Data Eng. Bull.},
 	pages = {39--45},
 	title = {{Accelerating the Machine Learning Lifecycle with MLflow}},
-	url = {https://api.semanticscholar.org/CorpusID:83459546},
+	url = {http://sites.computer.org/debull/A18dec/p39.pdf},
 	volume = {41},
 	year = {2018}
 }
