### 2025-12-15 9ec5fb6
Add clarification for LUMI venvs
