### 2025-12-15 3427f81
add dependencies description for other HPCs
Co-authored-by: Rakesh Sarma <126173968+r-sarma@users.noreply.github.com>
