+++ docs/installation/hpc_modules.rst
@@ -13,7 +13,7 @@ different HPC system, please refer to the *Other HPCs* tab.
 
     .. tab-item:: Juelich (JSC)
 
-        Modules for the `JUWELS <https://apps.fz-juelich.de/jsc/hps/juwels/configuration.html>`_ 
+        Modules for the `JUWELS <https://apps.fz-juelich.de/jsc/hps/juwels/configuration.html>`_
         system at Juelich Supercomputer (JSC):
 
         .. tab-set::
@@ -40,15 +40,15 @@ different HPC system, please refer to the *Other HPCs* tab.
 
     .. tab-item:: Vega
 
-        Modules for `Vega <https://doc.vega.izum.si/introduction/>`_ Supercomputer. 
-        
+        Modules for `Vega <https://doc.vega.izum.si/introduction/>`_ Supercomputer.
+
         When installing the environment from the login node, make sure that the
         CUDA drivers are loaded correctly and the GPU is visible by running the
         `nvidia-smi` command. This is very important for a successful
         installation of DeepSpeed and Horovod. If the GPU is not correctly
         visualized, consider logging-in again to another login node.
-        Alternatively, consider running the installation on a compute node. 
-        
+        Alternatively, consider running the installation on a compute node.
+
         .. tab-set::
 
             .. tab-item:: PyTorch
@@ -79,7 +79,7 @@ different HPC system, please refer to the *Other HPCs* tab.
                     ml Python/3.12.3-GCCcore-13.3.0
 
                     # Now you can create or active the python environment here
-        
+
         Currently, the latest version of ``mpi4py`` on Vega is not compatible with Python 3.12,
         therefore you'll have to build it yourself in your python environment:
 
@@ -107,6 +107,88 @@ different HPC system, please refer to the *Other HPCs* tab.
 
         These modules are needed to bind into the container the correct software suite on LUMI.
         More info can be found `here <https://lumi-supercomputer.github.io/LUMI-training-materials/ai-20250204/extra_05_RunningContainers/>`_.
+
+
+    .. tab-item:: Other HPCs
+
+        Module names and packaging conventions vary across HPC centres, but the underlying
+        software requirements for running ``itwinai`` are usually the same. The goal is to
+        load a consistent toolchain (compiler + MPI), a Python runtime, and (optionally) the
+        GPU communication stack.
+
+        The version numbers below are **reference values** known to work on at least one
+        modern HPC software stack; other versions may work as well.
+
+        **Recommended baseline (reference versions)**
+
+        * **Python**: >= 3.12 (e.g., 3.12.3)
+        * **CMake**: >= 3.29 (e.g., 3.29.3) — required to compile Horovod from source
+        * **git**: recent (e.g., 2.45.x)
+
+        **Compiler + MPI (ABI compatibility matters)**
+
+        * **GCC**: a recent major release (e.g., 13.3.0)
+        * **MPI**: a compatible MPI implementation (e.g., OpenMPI 5.0.x)
+        * **mpi4py**: built/installed against the *same* MPI you loaded (e.g., 4.0.1)
+
+        Many sites provide these as a single toolchain module (e.g., a compiler+MPI bundle).
+        That is fine as long as the compiler and MPI are internally compatible.
+
+        **GPU software stack (only if you use GPUs)**
+
+        * **CUDA toolkit/runtime**: >= 12.6 (e.g., 12.6)
+        * **cuDNN**: a CUDA-matched build (e.g., 9.5.0.* for CUDA 12)
+        * **NCCL**: recommended for multi-GPU communication (version typically follows CUDA)
+
+        Notes:
+
+        * For **PyTorch**, it is often possible to install wheels built for a CUDA version that
+          differs from the system CUDA. However, **DeepSpeed** is more sensitive: it generally
[…]
