### 2025-12-15 interTwin-eu/itwinai#468 @matbun
Thank you @r-sarma !

I have reworked your comment as an additional tab called "Other HPCs" in the installation guide.

#476 is now ready for discussion. @r-sarma please edit directly the code in the PR in case you want to make some changes.
