### 2025-12-15 3d8a597
Add a generic section in the installation document (#476)
* Update installation commands for GPU acceleration

* Update installation instructions for Horovod and DeepSpeed

* Add clarification for LUMI venvs

* add dependencies description for other HPCs

Co-authored-by: Rakesh Sarma <126173968+r-sarma@users.noreply.github.com>

* Add final disclaimer

* Improve formatting

---------

Co-authored-by: Rakesh Sarma <126173968+r-sarma@users.noreply.github.com>
