### 2025-12-18 #9409 @boisgera
> [@boisgera](https://github.com/boisgera) I am interested in reviewing this, can you please consider me?

Hi @Manishms18,

Thank you for this proposal!

Ideally, I'd rather not change the current team of reviewers at this stage in the process (the selection of reviewers is typically carried out [in the pre-review stage](https://github.com/openjournals/joss-reviews/issues?q=state%3Aopen%20label%3Apre-review)) and instead propose that you review future submissions (I've seen your profile in the reviewers database and I am definitely interested!)

However, if you are specifically interested in the itwinai project and you think you can carry out your review quickly, I can certainly make an exception.

What do you think?
