### 2025-12-18 #9409 @Manishms18
@boisgera Thank you for getting back to me. I completely understand and respect the usual process of selecting reviewers during the pre-review stage.

I’m happy to support future submissions as you suggested, and I’d really appreciate being considered for upcoming assignments.

Regarding itwinai specifically: if it would be helpful, I can take it on and complete the review quickly. I can deliver an initial review within 48–72 hours of assignment (and sooner if needed), and I’ll follow the journal’s review checklist and timelines closely.

Please let me know what works best on your side.
