### 2025-12-18 #9409 @xiazeyu
> Hi [@xiazeyu](https://github.com/xiazeyu),
>
> > Regarding to the Reproducibility, Functionality and Performance, unfortunately I am not able to confirm it.
> > I can only access Rivanna cluster, which do not have native support from itwinai. [...]
> > Rivanna has dropped support for singularity, and replaced that with apptainer. I do not know how to use apptainer to reproduce the results.
>
> If I understand the situation correctly: currently Rivanna is not officially supported by the itwinai project (at least in the version that was submitted to JOSS), but this is the platform that [@xiazeyu](https://github.com/xiazeyu) has easily access to (and therefore that they can use to evaluate the documented features of the software, which is required by JOSS).
>
> So AFAICT the best-case scenario is: by continuing to exchange informations you eventually manage to test itwinai (at least partially, maybe with rough edges) with Rivanna. But if this scenario doesn't pan out after a reasonable amount of time/effort, could you for example [@matbun](https://github.com/matbun) consider giving [@xiazeyu](https://github.com/xiazeyu) a temporary access to a platform that itwinai supports? (Any alternative, better idea to solve this is of course welcome!)

Thanks for following up. The major issue I experienced before was caused by testing on a document which actually pointed to the developing branch (instead of the stable one), and they prefer not to include that untested feature in the joss submission.

There's no issue related to the environment right now. And I think I can give some feedback in about 3 days. We can see alternative options after that.
