### 2025-12-18 #9409 @matbun
Hi @boisgera @xiazeyu ,
I will do my best to help @xiazeyu running itwinai on his cluster in the next days. I don't know how easy would be to give @xiazeyu access to JSC... @r-sarma is the best person for this, but tbh I would really leave this as the last resort!

I'm case we can quickly fix the problems with HPC access and deployment, do you think we'll be able to complete the review process by the end of December?

Thank you all for your help and feedback!
