### 2025-12-19 interTwin-eu/itwinai#481 @xiazeyu
"Required environment variable 'SLURM_NTASKS' not set"
I was trying the [MNIST example](https://github.com/interTwin-eu/itwinai/tree/main/use-cases/mnist/torch), and I ran into the following error:

```
(itwinai-dev) bash-4.4$itwinai exec-pipeline --config-name config.yaml +pipe_key=inference_pipeline
[INFO] itwinai.cli:956 - No steps selected. Executing the whole pipeline.
/sfs/gpfs/tardis/home/yxn7cj/itwinai-dev/.venv/lib/python3.11/site-packages/torch/cuda/__init__.py:61: FutureWarning: The pynvml package is deprecated. Please install nvidia-ml-py instead. If you did not install pynvml directly, please report this to the maintainers of the package that installed pynvml for you.
  import pynvml  # type: ignore[import]
Error executing job with overrides: ['+pipe_key=inference_pipeline']
Error in call to target 'itwinai.torch.inference.MulticlassTorchPredictor':
KeyError("Required environment variable 'SLURM_NTASKS' not set")
full_key: inference_pipeline.steps1

Set the environment variable HYDRA_FULL_ERROR=1 for a complete stack trace.
```

I asked Gemini and solved it by executing `export SLURM_NTASKS=1`. I think this is due to I was not running itwinai in a slurm environment.

Do you know why it might goes wrong for that? I did not assume a slurm environment is required to execute `itwinai exec-pipeline`.

ref https://github.com/openjournals/joss-reviews/issues/9409
