### 2025-12-19 #9409 @matbun
Thank you @xiazeyu for your feedback! Thanks for opening a new issue on the itwinai repo, so that we can track this.

@boisgera consider that both @xiazeyu and @yewentao256 seem to have completed their reviews, do you think that we can move to the next step?

Thank you!
