### 2025-12-19 16cf4c6
Enhance SLURM config with comments and fields
Added detailed comments and configuration fields for SLURM job.
