### 2025-12-19 f8754d0
Rename exp_name to experiment_name in YAML config
