### 2025-12-19 9cd56f5
Rename default pipe_key to 'training_pipeline'
