### 2025-12-19 1c8ae23
Rename default_pipe_key to 'training_pipeline'
