### 2025-12-19 e875707
Add configuration parameters to run-example.yaml
