### 2025-12-19 df3285d
Update config_path and config_name comments
Clarified config_path and config_name in YAML.
