+++ use-cases/mnist/torch/run-example.yaml
@@ -32,9 +32,10 @@ slurm_config:
   num_nodes: 2
   submit_job: true
   save_script: true
-  config_path: .
-  config_name: run-example
   pipe_key: training_pipeline
+  
+  config_name: run-example  # Assuming this is the name of this file
+  config_path: .  # Assuming that run-example.yaml is in the current directory
 
   pre_exec_file: https://raw.githubusercontent.com/interTwin-eu/itwinai/refs/heads/main/src/itwinai/slurm/system-base-scripts/vega_pre_exec.sh
 
