### 2025-12-19 interTwin-eu/itwinai#472 @matbun
@xiazeyu I have tried a fix on this branch: https://github.com/interTwin-eu/itwinai/tree/fix-itwinai-run-mnist

For instance, here you have the patched config file: https://github.com/interTwin-eu/itwinai/blob/fix-itwinai-run-mnist/use-cases/mnist/torch/run-example.yaml

I didnt have time to test it, but I think I understood where the problem was and I think I fixed it.. let's see :)
