### 2025-12-19 98037fc
Add experiment and run names to YAML configuration
