+++ use-cases/mnist/torch/run-example.yaml
@@ -33,6 +33,9 @@ slurm_config:
   submit_job: true
   save_script: true
   pipe_key: training_pipeline
+  experiment_name: "MNIST classifier"
+  run_name: "main-run"
+  dist_strat: ddp
   
   config_name: run-example  # Assuming this is the name of this file
   config_path: .  # Assuming that run-example.yaml is in the current directory
