### 2025-12-19 interTwin-eu/itwinai#472 @matbun
Could you try again? I have pushed a change to the branch adding

```yaml

  experiment_name: "MNIST classifier"
  run_name: "main-run"
  dist_strat: ddp

```

Again, not an expert of this chunk of code, but probably it fails when these fields are not provided...
