### 2025-12-19 a052610
Change dist_strat to use variable strategy
