### 2025-12-19 interTwin-eu/itwinai#472 @xiazeyu
> experiment_name: "MNIST classifier"
>   run_name: "main-run"
>   dist_strat: ${strategy}

I got `error: Option 'experiment_name' is not accepted` for this modification.

> experiment_name: "MNIST classifier"
> run_name: "main-run"

Got a `TypeError: TorchTrainer.__init__() got an unexpected keyword argument 'run_id'`.

After I commented out

```yaml
training_pipeline:
  steps:
    training_step:
     # run_id: ${run_id}
```

It's working! Note that I still need to `cd` into the `use-cases/mnist/torch` to use `itwinai run -jc run-example.yaml`.

I have the logs here, they seems to be good to me.

<summary>mnist-job.err</summary>

```
/sfs/gpfs/tardis/home/yxn7cj/itwinai-dev/.venv/lib/python3.11/site-packages/torch/cuda/__init__.py:61: FutureWarning: The pynvml package is deprecated. Please install nvidia-ml-py instead. If you did not install pynvml directly, please report this to the maintainers of the package that installed pynvml for you.
  import pynvml  # type: ignore[import]
/sfs/gpfs/tardis/home/yxn7cj/itwinai-dev/.venv/lib/python3.11/site-packages/torch/cuda/__init__.py:61: FutureWarning: The pynvml package is deprecated. Please install nvidia-ml-py instead. If you did not install pynvml directly, please report this to the maintainers of the package that installed pynvml for you.
  import pynvml  # type: ignore[import]
/sfs/gpfs/tardis/home/yxn7cj/itwinai-dev/.venv/lib/python3.11/site-packages/torch/cuda/__init__.py:61: FutureWarning: The pynvml package is deprecated. Please install nvidia-ml-py instead. If you did not install pynvml directly, please report this to the maintainers of the package that installed pynvml for you.
  import pynvml  # type: ignore[import]
/sfs/gpfs/tardis/home/yxn7cj/itwinai-dev/.venv/lib/python3.11/site-packages/torch/cuda/__init__.py:61: FutureWarning: The pynvml package is deprecated. Please install nvidia-ml-py instead. If you did not install pynvml directly, please report this to the maintainers of the package that installed pynvml for you.
  import pynvml  # type: ignore[import]
/sfs/gpfs/tardis/home/yxn7cj/itwinai-dev/.venv/lib/python3.11/site-packages/torch/cuda/__init__.py:61: FutureWarning: The pynvml package is deprecated. Please install nvidia-ml-py instead. If you did not install pynvml directly, please report this to the maintainers of the package that installed pynvml for you.
  import pynvml  # type: ignore[import]
/sfs/gpfs/tardis/home/yxn7cj/itwinai-dev/.venv/lib/python3.11/site-packages/torch/cuda/__init__.py:61: FutureWarning: The pynvml package is deprecated. Please install nvidia-ml-py instead. If you did not install pynvml directly, please report this to the maintainers of the package that installed pynvml for you.
  import pynvml  # type: ignore[import]
/sfs/gpfs/tardis/home/yxn7cj/itwinai-dev/.venv/lib/python3.11/site-packages/torch/cuda/__init__.py:61: FutureWarning: The pynvml package is deprecated. Please install nvidia-ml-py instead. If you did not install pynvml directly, please report this to the maintainers of the package that installed pynvml for you.
  import pynvml  # type: ignore[import]
/sfs/gpfs/tardis/home/yxn7cj/itwinai-dev/.venv/lib/python3.11/site-packages/torch/cuda/__init__.py:61: FutureWarning: The pynvml package is deprecated. Please install nvidia-ml-py instead. If you did not install pynvml directly, please report this to the maintainers of the package that installed pynvml for you.
  import pynvml  # type: ignore[import]
/sfs/gpfs/tardis/home/yxn7cj/itwinai-dev/.venv/lib/python3.11/site-packages/torch/cuda/__init__.py:61: FutureWarning: The pynvml package is deprecated. Please install nvidia-ml-py instead. If you did not install pynvml directly, please report this to the maintainers of the package that installed pynvml for you.
  import pynvml  # type: ignore[import]
/sfs/gpfs/tardis/home/yxn7cj/itwinai-dev/.venv/lib/python3.11/site-packages/torch/cuda/__init__.py:61: FutureWarning: The pynvml package is deprecated. Please install nvidia-ml-py instead. If you did not install pynvml directly, please report this to the maintainers of the package that installed pynvml for you.
  import pynvml  # type: ignore[import]
/sfs/gpfs/tardis/home/yxn7cj/itwinai-dev/.venv/lib/python3.11/site-packages/torch/cuda/__init__.py:61: FutureWarning: The pynvml package is deprecated. Please install nvidia-ml-py instead. If you did not install pynvml directly, please report this to the maintainers of the package that installed pynvml for you.
  import pynvml  # type: ignore[import]
/sfs/gpfs/tardis/home/yxn7cj/itwinai-dev/.venv/lib/python3.11/site-packages/torch/cuda/__init__.py:61: FutureWarning: The pynvml package is deprecated. Please install nvidia-ml-py instead. If you did not install pynvml directly, please report this to the maintainers of the package that installed pynvml for you.
  import pynvml  # type: ignore[import]
/sfs/gpfs/tardis/home/yxn7cj/itwinai-dev/.venv/lib/python3.11/site-packages/mlflow/tracking/_tracking_service/utils.py:177: FutureWarning: The filesystem tracking backend (e.g., './mlruns') will be deprecated in February 2026. Consider transitioning to a database backend (e.g., 'sqlite:///mlflow.db') to take advantage of the latest MLflow features. See https://github.com/mlflow/mlflow/issues/18534 for more details and migration guidance.
  return FileStore(store_uri, store_uri)
2025/12/19 17:18:41 INFO mlflow.tracking.fluent: Experiment with name 'MNIST classifier' does not exist. Creating a new experiment.
/sfs/gpfs/tardis/home/yxn7cj/itwinai-dev/.venv/lib/python3.11/site-packages/mlflow/tracking/_tracking_service/utils.py:177: FutureWarning: The filesystem tracking backend (e.g., './mlruns') will be deprecated in February 2026. Consider transitioning to a database backend (e.g., 'sqlite:///mlflow.db') to take advantage of the latest MLflow features. See https://github.com/mlflow/mlflow/issues/18534 for more details and migration guidance.
  return FileStore(store_uri, store_uri)
/sfs/gpfs/tardis/home/yxn7cj/itwinai-dev/.venv/lib/python3.11/site-packages/mlflow/tracking/_tracking_service/utils.py:177: FutureWarning: The filesystem tracking backend (e.g., './mlruns') will be deprecated in February 2026. Consider transitioning to a database backend (e.g., 'sqlite:///mlflow.db') to take advantage of the latest MLflow features. See https://github.com/mlflow/mlflow/issues/18534 for more details and migration guidance.
  return FileStore(store_uri, store_uri)
/sfs/gpfs/tardis/home/yxn7cj/itwinai-dev/.venv/lib/python3.11/site-packages/mlflow/tracking/_tracking_service/utils.py:177: FutureWarning: The filesystem tracking backend (e.g., './mlruns') will be deprecated in February 2026. Consider transitioning to a database backend (e.g., 'sqlite:///mlflow.db') to take advantage of the latest MLflow features. See https://github.com/mlflow/mlflow/issues/18534 for more details and migration guidance.
  return FileStore(store_uri, store_uri)
/sfs/gpfs/tardis/home/yxn7cj/itwinai-dev/.venv/lib/python3.11/site-packages/mlflow/tracking/_tracking_service/utils.py:177: FutureWarning: The filesystem tracking backend (e.g., './mlruns') will be deprecated in February 2026. Consider transitioning to a database backend (e.g., 'sqlite:///mlflow.db') to take advantage of the latest MLflow features. See https://github.com/mlflow/mlflow/issues/18534 for more details and migration guidance.
  return FileStore(store_uri, store_uri)
/sfs/gpfs/tardis/home/yxn7cj/itwinai-dev/.venv/lib/python3.11/site-packages/mlflow/tracking/_tracking_service/utils.py:177: FutureWarning: The filesystem tracking backend (e.g., './mlruns') will be deprecated in February 2026. Consider transitioning to a database backend (e.g., 'sqlite:///mlflow.db') to take advantage of the latest MLflow features. See https://github.com/mlflow/mlflow/issues/18534 for more details and migration guidance.
  return FileStore(store_uri, store_uri)
/sfs/gpfs/tardis/home/yxn7cj/itwinai-dev/.venv/lib/python3.11/site-packages/mlflow/tracking/_tracking_service/utils.py:177: FutureWarning: The filesystem tracking backend (e.g., './mlruns') will be deprecated in February 2026. Consider transitioning to a database backend (e.g., 'sqlite:///mlflow.db') to take advantage of the latest MLflow features. See https://github.com/mlflow/mlflow/issues/18534 for more details and migration guidance.
  return FileStore(store_uri, store_uri)
/sfs/gpfs/tardis/home/yxn7cj/itwinai-dev/.venv/lib/python3.11/site-packages/mlflow/tracking/_tracking_service/utils.py:177: FutureWarning: The filesystem tracking backend (e.g., './mlruns') will be deprecated in February 2026. Consider transitioning to a database backend (e.g., 'sqlite:///mlflow.db') to take advantage of the latest MLflow features. See https://github.com/mlflow/mlflow/issues/18534 for more details and migration guidance.
  return FileStore(store_uri, store_uri)

Epochs:   0%|          | 0/5 [00:00<?, ?it/s]
Epoch 1/5:   0%|          | 0/5 [00:00<?, ?it/s]

Train batches:   0%|          | 0/7 [00:00<?, ?it/s][A

Train batches:  14%|█▍        | 1/7 [00:01<00:07,  1.26s/it][A

Train batches: 9it [00:01,  8.84it/s]                       [A

Train batches: 18it [00:01, 18.70it/s][A

Train batches: 27it [00:01, 28.99it/s][A

Train batches: 36it [00:01, 39.17it/s][A

Train batches: 45it [00:01, 48.94it/s][A

Train batches: 54it [00:01, 56.27it/s][A

                                      [A

Validation batches:   0%|          | 0/1 [00:00<?, ?it/s][A

                                                         [A
Epoch 1/5:  20%|██        | 1/5 [00:03<00:13,  3.49s/it]
Epoch 2/5:  20%|██        | 1/5 [00:03<00:13,  3.49s/it]

Train batches:   0%|          | 0/7 [00:00<?, ?it/s][A

Train batches:  71%|███████▏  | 5/7 [00:00<00:00, 44.98it/s][A

Train batches: 11it [00:00, 51.79it/s]                      [A

Train batches: 19it [00:00, 59.31it/s][A

Train batches: 27it [00:00, 64.94it/s][A

Train batches: 35it [00:00, 69.37it/s][A

Train batches: 43it [00:00, 72.79it/s][A

Train batches: 51it [00:00, 73.72it/s][A

                                      [A

Validation batches:   0%|          | 0/1 [00:00<?, ?it/s][A

                                                         [A
Epoch 2/5:  40%|████      | 2/5 [00:04<00:06,  2.19s/it]
Epoch 3/5:  40%|████      | 2/5 [00:04<00:06,  2.19s/it]

Train batches:   0%|          | 0/7 [00:00<?, ?it/s][A

Train batches:  86%|████████▌ | 6/7 [00:00<00:00, 55.58it/s][A

Train batches: 16it [00:00, 75.32it/s]                      [A

Train batches: 25it [00:00, 77.90it/s][A

Train batches: 35it [00:00, 81.57it/s][A

Train batches: 44it [00:00, 83.49it/s][A

Train batches: 53it [00:00, 83.72it/s][A

                                      [A

Validation batches:   0%|          | 0/1 [00:00<?, ?it/s][A

                                                         [A
Epoch 3/5:  60%|██████    | 3/5 [00:05<00:03,  1.72s/it]
Epoch 4/5:  60%|██████    | 3/5 [00:05<00:03,  1.72s/it]

Train batches:   0%|          | 0/7 [00:00<?, ?it/s][A

Train batches:  71%|███████▏  | 5/7 [00:00<00:00, 41.09it/s][A

Train batches: 13it [00:00, 61.10it/s]                      [A

Train batches: 21it [00:00, 66.07it/s][A

Train batches: 31it [00:00, 75.20it/s][A

Train batches: 40it [00:00, 79.19it/s][A

Train batches: 48it [00:00, 76.97it/s][A

Train batches: 56it [00:00, 77.19it/s][A

                                      [A

Validation batches:   0%|          | 0/1 [00:00<?, ?it/s][A

                                                         [A
Epoch 4/5:  80%|████████  | 4/5 [00:07<00:01,  1.54s/it]
Epoch 5/5:  80%|████████  | 4/5 [00:07<00:01,  1.54s/it]

Train batches:   0%|          | 0/7 [00:00<?, ?it/s][A

Train batches:  86%|████████▌ | 6/7 [00:00<00:00, 59.71it/s][A

Train batches: 13it [00:00, 64.98it/s]                      [A

Train batches: 22it [00:00, 73.22it/s][A

Train batches: 30it [00:00, 74.84it/s][A

Train batches: 39it [00:00, 78.28it/s][A

Train batches: 47it [00:00, 76.08it/s][A

Train batches: 59it [00:00, 89.18it/s][A

                                      [A

Validation batches:   0%|          | 0/1 [00:00<?, ?it/s][A

                                                         [A
Epoch 5/5: 100%|██████████| 5/5 [00:08<00:00,  1.41s/it]
Epoch 5/5: 100%|██████████| 5/5 [00:08<00:00,  1.67s/it]

```

<summary>mnist-job.out</summary>

```
DEBUG: SLURM_SUBMIT_DIR: /sfs/gpfs/tardis/home/yxn7cj/itwinai-dev/use-cases/mnist/torch
DEBUG: SLURM_JOB_ID: 6779169
DEBUG: SLURM_JOB_NODELIST: udc-aj36-35,udc-aj38-35
DEBUG: SLURM_JOB_NUM_NODES: 2
DEBUG: SLURM_NTASKS: 2
DEBUG: SLURM_TASKS_PER_NODE: 1(x2)
DEBUG: SLURM_SUBMIT_HOST: udc-aj36-36
DEBUG: SLURMD_NODENAME: udc-aj36-35
DEBUG: SLURM_GPUS_PER_NODE: 4
DEBUG: CUDA_VISIBLE_DEVICES (before): 0,1,2,3
DEBUG: DIST_MODE:

 Usage: itwinai [OPTIONS] COMMAND [ARGS]...

 itwinai command line interface.

╭─ Options ────────────────────────────────────────────────────────────────────╮
│ --version             -v        Show version and exit.                       │
│ --install-completion            Install completion for the current shell.    │
│ --show-completion               Show completion for the current shell, to    │
│                                 copy it or customize the installation.       │
│ --help                          Show this message and exit.                  │
╰──────────────────────────────────────────────────────────────────────────────╯
╭─ Commands ───────────────────────────────────────────────────────────────────╮
│ generate-flamegraph           Generates a flamegraph from the given          │
│                               profiling output.                              │
│ generate-py-spy-report        Generates an aggregation of the raw py-spy     │
│                               profiling data, showing which leaf functions   │
│                               collected the most samples.                    │
│ generate-scalability-report   Generates scalability reports for epoch time,  │
│                               GPU data, and communication data based the     │
│                               mlflow logs.                                   │
│ sanity-check                  Run sanity checks on the installation of       │
│                               itwinai and its dependencies by trying to      │
│                               import itwinai modules. By default, only       │
│                               itwinai core modules (neither torch, nor       │
│                               tensorflow) are tested.                        │
│ check-distributed-cluster     This command provides a suite of tests for a   │
│                               quick sanity check of the network setup for    │
│                               torch distributed. Useful when working with    │
│                               containers on HPC. Remember to prepend         │
│                               *torchrun* in front of this command or to      │
│                               start a *Ray* cluster.                         │
│ generate-slurm                Generates a default SLURM script using         │
│                               arguments and optionally a configuration file. │
│ run                           Launch ML jobs with dependency installation    │
│                               and SLURM scheduling.                          │
│ exec-pipeline                 Execute a pipeline from configuration file     │
│                               using Hydra CLI. Allows dynamic override of    │
│                               fields which can be appended as a list of      │
│                               overrides (e.g., batch_size=32). By default,   │
│                               it will expect a configuration file called     │
│                               "config.yaml" in the current working           │
│                               directory. To override the default behavior    │
│                               set --config-name and --config-path. By        │
│                               default, this command will execute the whole   │
│                               pipeline under "training_pipeline" field in    │
│                               the configuration file. To execute a different │
│                               pipeline you can override this by passing      │
│                               "+pipe_key=your_pipeline" in the list of       │
│                               overrides, and to execute only a subset of the │
│                               steps, you can pass "+pipe_steps=[0,1]".       │
│ mlflow-ui                     Visualize logs with Mlflow.                    │
│ mlflow-server                 Spawn Mlflow server.                           │
│ kill-mlflow-server            Kill Mlflow server.                            │
│ download-mlflow-data          Download metrics data from MLFlow experiments  │
│                               and save to a CSV file.                        │
│ tensorboard-ui                Visualize logs with TensorBoard.               │
╰──────────────────────────────────────────────────────────────────────────────╯

usage: torchrun [-h] [--nnodes NNODES] [--nproc-per-node NPROC_PER_NODE]
                [--rdzv-backend RDZV_BACKEND] [--rdzv-endpoint RDZV_ENDPOINT]
                [--rdzv-id RDZV_ID] [--rdzv-conf RDZV_CONF] [--standalone]
                [--max-restarts MAX_RESTARTS]
                [--monitor-interval MONITOR_INTERVAL]
                [--start-method {spawn,fork,forkserver}] [--role ROLE] [-m]
                [--no-python] [--run-path] [--log-dir LOG_DIR] [-r REDIRECTS]
                [-t TEE] [--local-ranks-filter LOCAL_RANKS_FILTER]
                [--node-rank NODE_RANK] [--master-addr MASTER_ADDR]
                [--master-port MASTER_PORT] [--local-addr LOCAL_ADDR]
                [--logs-specs LOGS_SPECS]
                training_script ...

Torch Distributed Elastic Training Launcher

positional arguments:
  training_script       Full path to the (single GPU) training program/script
                        to be launched in parallel, followed by all the
                        arguments for the training script.
  training_script_args

options:
  -h, --help            show this help message and exit
  --nnodes NNODES       Number of nodes, or the range of nodes in form
                        <minimum_nodes>:<maximum_nodes>.
  --nproc-per-node NPROC_PER_NODE, --nproc_per_node NPROC_PER_NODE
                        Number of workers per node; supported values: [auto,
                        cpu, gpu, int].
  --rdzv-backend RDZV_BACKEND, --rdzv_backend RDZV_BACKEND
                        Rendezvous backend.
  --rdzv-endpoint RDZV_ENDPOINT, --rdzv_endpoint RDZV_ENDPOINT
                        Rendezvous backend endpoint; usually in form
                        <host>:<port>.
  --rdzv-id RDZV_ID, --rdzv_id RDZV_ID
                        User-defined group id.
  --rdzv-conf RDZV_CONF, --rdzv_conf RDZV_CONF
                        Additional rendezvous configuration
                        (<key1>=<value1>,<key2>=<value2>,...).
  --standalone          Start a local standalone rendezvous backend that is
                        represented by a C10d TCP store on a free port. Useful
                        when launching single-node, multi-worker job. If
                        specified --rdzv-backend, --rdzv-endpoint, --rdzv-id
                        are auto-assigned and any explicitly set values are
                        ignored.
  --max-restarts MAX_RESTARTS, --max_restarts MAX_RESTARTS
                        Maximum number of worker group restarts before
                        failing.
  --monitor-interval MONITOR_INTERVAL, --monitor_interval MONITOR_INTERVAL
                        Interval, in seconds, to monitor the state of workers.
  --start-method {spawn,fork,forkserver}, --start_method {spawn,fork,forkserver}
[…]
