### 2025-12-21 4525dcf
Delete src/itwinai/slurm/sample_slurm_config.yaml
