### 2025-12-21 3236fbe
Add missing modules to Python SDK docs
