### 2025-12-21 a03fdb0
Draft refactor slurm builder
