+++ src/itwinai/slurm/slurm_script_builder.py
@@ -1,517 +1,0 @@
-# --------------------------------------------------------------------------------------
-# Part of the interTwin Project: https://www.intertwin.eu/
-#
-# Created by: Jarl Sondre Sæther
-#
-# Credit:
-# - Jarl Sondre Sæther <jarl.sondre.saether@cern.ch> - CERN
-# - Matteo Bunino <matteo.bunino@cern.ch> - CERN
-# --------------------------------------------------------------------------------------
-
-import logging
-import subprocess
-from collections.abc import Iterable
-from pathlib import Path
-from tempfile import NamedTemporaryFile
-from typing import Dict, List, Literal
-
-import typer
-from requests.exceptions import RequestException
-from validators import url
-
-from itwinai.slurm.slurm_constants import (
-    DEFAULT_PY_SPY_DIR,
-    DEFAULT_SLURM_LOG_DIR,
-    DEFAULT_SLURM_SAVE_DIR,
-)
-from itwinai.slurm.slurm_script_configuration import SlurmScriptConfiguration
-from itwinai.slurm.utils import (
-    get_slurm_job_parser,
-    remove_indentation_from_multiline_string,
-    retrieve_remote_file,
-)
-
-cli_logger = logging.getLogger("cli_logger")
-
-
-class SlurmScriptBuilder:
-    def __init__(
-        self,
-        slurm_script_configuration: SlurmScriptConfiguration,
-        should_submit: bool,
-        should_save: bool,
-        save_dir: str | Path | None = None,
-        pre_exec_file: str | None = None,
-        exec_file: str | None = None,
-    ):
-        self.slurm_script_configuration = slurm_script_configuration
-        self.should_submit = should_submit
-        self.should_save = should_save
-        self.pre_exec_file = pre_exec_file
-        self.exec_file = exec_file
-
-        if isinstance(save_dir, str):
-            save_dir = Path(save_dir)
-        self.save_dir = save_dir
-
-    @staticmethod
-    def submit_script(script: str) -> None:
-        """Submits the given script with 'sbatch' using a temporary file."""
-        with NamedTemporaryFile(mode="w") as temp_file:
-            temp_file.write(script)
-            # Making sure the script is written to file before the cmd is launched
-            temp_file.flush()
-            try:
-                subprocess.run(["sbatch", temp_file.name], check=True)
-            except FileNotFoundError as e:
-                cli_logger.error(
-                    "'sbatch' failed. Are you sure you have it installed on your system? Error"
-                    f" message:\n{e}."
-                )
-                raise typer.Exit(1)
-            except subprocess.CalledProcessError as e:
-                cli_logger.error(f"'sbatch' failed to submit script. Error message:\n{e}")
-                raise typer.Exit(1)
-
-    @staticmethod
-    def save_script(script: str, file_path: Path) -> None:
-        """Saves the given script to the given file path."""
-        if not file_path.parent.exists():
-            cli_logger.info(f"Creating directory '{file_path.parent.resolve()}'!")
-            file_path.parent.mkdir(parents=True, exist_ok=True)
-
-        try:
-            with open(file_path, "x") as f:
-                f.write(script)
-            cli_logger.info(f"Saved SLURM script to '{file_path.resolve()}'.")
-        except FileExistsError:
-            cli_logger.error(
-                f"File '{file_path.resolve()}' already exists! Give a different path or "
-                "delete the file first."
-            )
-            raise typer.Exit(1)
-
-    @staticmethod
-    def print_script(script: str) -> None:
-        """Prints the given script to stdout using the cli_logger."""
-        upper_banner_str = f"{'#' * 20} SLURM Script Preview {'#' * 20}"
[…]
