+++ src/itwinai/slurm/configuration.py
@@ -1,0 +1,195 @@
+# --------------------------------------------------------------------------------------
+# Part of the interTwin Project: https://www.intertwin.eu/
+#
+# Created by: Jarl Sondre Sæther
+#
+# Credit:
+# - Jarl Sondre Sæther <jarl.sondre.saether@cern.ch> - CERN
+# - Matteo Bunino <matteo.bunino@cern.ch> - CERN
+# --------------------------------------------------------------------------------------
+
+from pathlib import Path
+from typing import List, Literal
+
+from pydantic import BaseModel, ConfigDict, Field, field_validator
+
+from .constants import SLURM_TEMPLATE, DEFAULT_SLURM_SAVE_DIR
+from .utils import remove_indentation_from_multiline_string
+
+DEFAULT_TRAINING_CMD = (
+    "{itwinai_launcher} exec-pipeline "
+    "--config-name={config_name} "
+    "--config-path={config_path} "
+    "--strategy={distributed_strategy} "
+    "--run_name={run_name}"
+    "+pipe_key={pipe_key} "
+)
+
+
+class SlurmScriptConfiguration(BaseModel):
+    """Configuration object for the SLURM script. It contains all the settings for the
+    SLURM script such as which hardware you are requesting or for how long to run it.
+    As it allows for any ``pre_exec_command`` and ``exec_command``, it should work for
+    any SLURM script.
+    """
+
+    model_config = ConfigDict(extra="forbid")
+
+    #: Optional job name for the SLURM job.
+    job_name: str | None = None
+    #: Billing account to charge the job to.
+    account: str = "intertwin"
+    #: Partition/queue the job should run on.
+    partition: str = "develbooster"
+    #: Wall-clock time limit for the job (HH:MM:SS).
+    time: str = "00:30:00"
+
+    #: Path to standard output file.
+    std_out: Path | None = None
+    #: Path to standard error file.
+    err_out: Path | None = None
+
+    #: Number of nodes requested.
+    num_nodes: int = 1
+    #: Number of tasks per node.
+    num_tasks_per_node: int = 1
+    #: GPUs per node requested.
+    gpus_per_node: int = 4
+    #: CPUs per task requested.
+    cpus_per_task: int = 16
+    #: Memory per node requested.
+    memory: str = "16G"
+    #: Whether to request exclusive node access.
+    exclusive: bool = False
+
+    #: Pre-execution command content (shell).
+    #: Typically used to set up the environment before executing the command,
+    #: e.g. "ml Python", "source .venv/bin/activate" etc.
+    #: Usually this should not be set by the user except for advanced use cases, and it
+    #: will be generated by the SLURM script builder based on the configuration.
+    pre_exec_command: str | None = None
+
+    #: Main execution command content (shell).
+    #: Command to execute, typically an 'srun' command.
+    #: Usually this should not be set by the user except for advanced use cases, and it
+    #: will be generated by the SLURM script builder based on the configuration.
+    exec_command: str | None = None
+
+    # Builder behavior
+    #: Whether to save the generated SLURM script.
+    save_script: bool = False
+    #: Whether to submit the generated SLURM script.
+    submit_job: bool = False
+    #: Directory where the script should be saved.
+    save_dir: Path | None = Field(
+        default=Path(DEFAULT_SLURM_SAVE_DIR),
+        description="Directory to write the generated SLURM script.",
+    )
+    #: Path/URL to a pre-exec file to load content from.
+    pre_exec_file: str | None = None
+    #: Path/URL to an exec file to load content from.
+    exec_file: str | None = None
+
+    def exclusive_line(self) -> str:
+        return "#SBATCH --exclusive" if self.exclusive else ""
+
+    def generate_script(self) -> str:
+        """Uses the provided configuration parameters and formats a SLURM script with
+        the requested settings.
[…]
