+++ use-cases/mnist/torch/slurm_config.yaml
@@ -4,13 +4,13 @@ python_venv: ../../../.venv/
 account: s24r05-03-users
 partition: gpu
 
-dist_strat: ddp
+distributed_strategy: ddp
 pipe_key: training_pipeline
 
 py_spy: false
 profiling_sampling_rate: 100
 
 training_cmd: "$(which itwinai) exec-pipeline \
-  strategy={dist_strat} \
-  checkpoints_location=checkpoints_{dist_strat} \
+  strategy={distributed_strategy} \
+  checkpoints_location=checkpoints_{distributed_strategy} \
   +pipe_key={pipe_key}"

[…]
