### 2025-12-21 d2a5c07
Fix linter and docs
