+++ src/itwinai/slurm/configuration.py
@@ -13,7 +13,7 @@ from typing import List, Literal
 
 from pydantic import BaseModel, ConfigDict, Field, field_validator
 
-from .constants import SLURM_TEMPLATE, DEFAULT_SLURM_SAVE_DIR
+from .constants import DEFAULT_SLURM_SAVE_DIR, SLURM_TEMPLATE
 from .utils import remove_indentation_from_multiline_string
 
 DEFAULT_TRAINING_CMD = (
@@ -21,7 +21,7 @@ DEFAULT_TRAINING_CMD = (
     "--config-name={config_name} "
     "--config-path={config_path} "
     "--strategy={distributed_strategy} "
-    "--run_name={run_name}"
+    "--run_name={run_name} "
     "+pipe_key={pipe_key} "
 )
 
@@ -35,59 +35,60 @@ class SlurmScriptConfiguration(BaseModel):
 
     model_config = ConfigDict(extra="forbid")
 
-    #: Optional job name for the SLURM job.
+    #: Optional job name for the SLURM job. Default: None (auto-generated later).
     job_name: str | None = None
-    #: Billing account to charge the job to.
+    #: Billing account to charge the job to. Default: "intertwin".
     account: str = "intertwin"
-    #: Partition/queue the job should run on.
+    #: Partition/queue the job should run on. Default: "develbooster".
     partition: str = "develbooster"
-    #: Wall-clock time limit for the job (HH:MM:SS).
+    #: Wall-clock time limit for the job (HH:MM:SS). Default: "00:30:00".
     time: str = "00:30:00"
 
-    #: Path to standard output file.
+    #: Path to standard output file. Default: None (filled later).
     std_out: Path | None = None
-    #: Path to standard error file.
+    #: Path to standard error file. Default: None (filled later).
     err_out: Path | None = None
 
-    #: Number of nodes requested.
+    #: Number of nodes requested. Default: 1.
     num_nodes: int = 1
-    #: Number of tasks per node.
+    #: Number of tasks per node. Default: 1.
     num_tasks_per_node: int = 1
-    #: GPUs per node requested.
+    #: GPUs per node requested. Default: 4.
     gpus_per_node: int = 4
-    #: CPUs per task requested.
+    #: CPUs per task requested. Default: 16.
     cpus_per_task: int = 16
-    #: Memory per node requested.
+    #: Memory per node requested. Default: "16G".
     memory: str = "16G"
-    #: Whether to request exclusive node access.
+    #: Whether to request exclusive node access. Default: False.
     exclusive: bool = False
 
-    #: Pre-execution command content (shell).
+    #: Pre-execution command content (shell). Default: None (set by builder).
     #: Typically used to set up the environment before executing the command,
     #: e.g. "ml Python", "source .venv/bin/activate" etc.
     #: Usually this should not be set by the user except for advanced use cases, and it
     #: will be generated by the SLURM script builder based on the configuration.
     pre_exec_command: str | None = None
 
-    #: Main execution command content (shell).
+    #: Main execution command content (shell). Default: None (set by builder).
     #: Command to execute, typically an 'srun' command.
     #: Usually this should not be set by the user except for advanced use cases, and it
     #: will be generated by the SLURM script builder based on the configuration.
     exec_command: str | None = None
 
     # Builder behavior
-    #: Whether to save the generated SLURM script.
+    #: Whether to save the generated SLURM script. Default: False.
     save_script: bool = False
-    #: Whether to submit the generated SLURM script.
+    #: Whether to submit the generated SLURM script. Default: False.
     submit_job: bool = False
-    #: Directory where the script should be saved.
[…]
