+++ use-cases/mnist/torch/run-example.yaml
@@ -53,7 +53,7 @@ slurm_config:
     "--config-name={config_name} "
     "--config-path={config_path} "
     "--strategy={distributed_strategy} "
-    "--run_name={run_name}"
+    "--run_name={run_name} "
     "+pipe_key={pipe_key} "
 
 # Workflows configuration
