### 2025-12-21 d0ba193
Fix exec-pipeline and small typos
