+++ src/itwinai/slurm/configuration.py
@@ -21,7 +21,7 @@ DEFAULT_TRAINING_CMD = (
     "--config-name={config_name} "
     "--config-path={config_path} "
     "--strategy={distributed_strategy} "
-    "--run_name={run_name} "
+    "--run-name={run_name} "
     "+pipe_key={pipe_key} "
 )
 
@@ -35,60 +35,60 @@ class SlurmScriptConfiguration(BaseModel):
 
     model_config = ConfigDict(extra="forbid")
 
-    #: Optional job name for the SLURM job. Default: None (auto-generated later).
+    #: Optional job name for the SLURM job. Defaults to None (auto-generated later).
     job_name: str | None = None
-    #: Billing account to charge the job to. Default: "intertwin".
+    #: Billing account to charge the job to. Defaults to "intertwin".
     account: str = "intertwin"
-    #: Partition/queue the job should run on. Default: "develbooster".
+    #: Partition/queue the job should run on. Defaults to "develbooster".
     partition: str = "develbooster"
-    #: Wall-clock time limit for the job (HH:MM:SS). Default: "00:30:00".
+    #: Wall-clock time limit for the job (HH:MM:SS). Defaults to "00:30:00".
     time: str = "00:30:00"
 
-    #: Path to standard output file. Default: None (filled later).
+    #: Path to standard output file. Defaults to None (filled later).
     std_out: Path | None = None
-    #: Path to standard error file. Default: None (filled later).
+    #: Path to standard error file. Defaults to None (filled later).
     err_out: Path | None = None
 
-    #: Number of nodes requested. Default: 1.
+    #: Number of nodes requested. Defaults to 1.
     num_nodes: int = 1
-    #: Number of tasks per node. Default: 1.
+    #: Number of tasks per node. Defaults to 1.
     num_tasks_per_node: int = 1
-    #: GPUs per node requested. Default: 4.
+    #: GPUs per node requested. Defaults to 4.
     gpus_per_node: int = 4
-    #: CPUs per task requested. Default: 16.
+    #: CPUs per task requested. Defaults to 16.
     cpus_per_task: int = 16
-    #: Memory per node requested. Default: "16G".
+    #: Memory per node requested. Defaults to "16G".
     memory: str = "16G"
-    #: Whether to request exclusive node access. Default: False.
+    #: Whether to request exclusive node access. Defaults to False.
     exclusive: bool = False
 
-    #: Pre-execution command content (shell). Default: None (set by builder).
+    #: Pre-execution command content (shell). Defaults to None (set by builder).
     #: Typically used to set up the environment before executing the command,
     #: e.g. "ml Python", "source .venv/bin/activate" etc.
     #: Usually this should not be set by the user except for advanced use cases, and it
     #: will be generated by the SLURM script builder based on the configuration.
     pre_exec_command: str | None = None
 
-    #: Main execution command content (shell). Default: None (set by builder).
+    #: Main execution command content (shell). Defaults to None (set by builder).
     #: Command to execute, typically an 'srun' command.
     #: Usually this should not be set by the user except for advanced use cases, and it
     #: will be generated by the SLURM script builder based on the configuration.
     exec_command: str | None = None
 
     # Builder behavior
-    #: Whether to save the generated SLURM script. Default: False.
+    #: Whether to save the generated SLURM script. Defaults to False.
     save_script: bool = False
-    #: Whether to submit the generated SLURM script. Default: False.
+    #: Whether to submit the generated SLURM script. Defaults to False.
     submit_job: bool = False
[…]
