+++ docs/getting-started/complete-workflow-example.rst
@@ -125,55 +125,87 @@ using the itwinai MNIST plugin. This simplified example shows the key components
 
 .. code-block:: yaml
 
-   # General training parameters
-   run_id: mnist-usecase-0
-   dataset_root: .tmp/
-   batch_size: 128
-   lr: 0.001
-   epochs: 5
-   strategy: ddp
-   # The rest of the general config is omitted for the sake of readability
+    # Default fields (always needed)
+    strategy: ddp
+    run_name: "mnist"
 
-   # Install the itwinai MNIST plugin from GitHub
-   plugins:
-     - git+https://github.com/matbun/itwinai-mnist-plugin.git
+    # General config
+    dataset_root: .tmp/
+    num_classes: 10
+    batch_size: 128
+    num_workers_dataloader: 4
+    pin_memory: False
+    lr: 0.001
+    momentum: 0.9
+    fp16_allreduce: False
+    use_adasum: False
+    gradient_predivide_factor: 1.0
+    epochs: 5
+    test_data_path: mnist-sample-data
+    inference_model_mlflow_uri: mnist-pre-trained.pth
+    predictions_dir: mnist-predictions
+    predictions_file: predictions.csv
+    class_labels: null
+    checkpoints_location: checkpoints
+    checkpoint_every: 1
 
-   # SLURM configuration for Vega HPC system
-   slurm_config:
-     job_name: mnist-job
-     account: s24r05-03-users  # Replace with your account
-     partition: gpu
-     memory: 64G
-     num_nodes: 2
-     submit_job: true
-     save_script: true
-     pre_exec_file: https://raw.githubusercontent.com/interTwin-eu/itwinai/refs/heads/main/src/itwinai/slurm/system-base-scripts/vega_pre_exec.sh
+    plugins:
+      - git+https://github.com/matbun/itwinai-mnist-plugin.git
 
-   # Define the ML pipeline
-   training_pipeline:
-     _target_: itwinai.pipeline.Pipeline
-     steps:
-       dataloading_step:
-         _target_: itwinai.plugins.mnist.dataloader.MNISTDataModuleTorch
-         save_path: ${dataset_root}
-       training_step:
-         _target_: itwinai.torch.trainer.TorchTrainer
-         run_id: ${run_id}
-         config:
-           batch_size: ${batch_size}
-           optimizer: sgd
-           optim_lr: ${lr}
-         model:
-           _target_: itwinai.plugins.mnist.model.Net
-         epochs: ${epochs}
-         strategy: ${strategy}
-         logger:
-           _target_: itwinai.loggers.LoggersCollection
-           loggers:
-             - _target_: itwinai.loggers.ConsoleLogger
-             - _target_: itwinai.loggers.MLFlowLogger
-               experiment_name: MNIST classifier
-       ... # The rest of the pipeline is omitted for the sake of readability
+    slurm_config:
+      job_name: mnist-job
+      account: s24r05-03-users
+      partition: gpu
+      memory: 64G
+      mode: single
+      num_nodes: 2
+      submit_job: true
+      save_script: true
+      pre_exec_file: https://raw.githubusercontent.com/interTwin-eu/itwinai/refs/heads/main/src/itwinai/slurm/system-base-scripts/vega_pre_exec.sh
+
+      # Fields referring to this config file
+      pipe_key: training_pipeline
+      config_name: run-example # Assuming this is the name of this file
+      config_path: . # Assuming that run-example.yaml is in the current directory
+
+      # Propagate global config to this slurm_config
+      distributed_strategy: ${strategy}
+      run_name: ${run_name}
+
+      # This provides a template for the training command launched
+      # using itwinai exec-pipeline -c <this config>.yaml
+      # The template is filled in using the fields of this slurm_config.
[…]
