### 2025-12-21 a1ba827
Remove unnecessary default values
