+++ src/itwinai/slurm/constants.py
@@ -24,6 +24,7 @@ SLURM_TEMPLATE = r"""#!/bin/bash
 
 # Resource allocation
 #SBATCH --nodes={num_nodes}
+#SBATCH --ntasks={num_tasks}
 #SBATCH --ntasks-per-node={num_tasks_per_node}
 #SBATCH --cpus-per-task={cpus_per_task}
 #SBATCH --gpus-per-node={gpus_per_node}
