### 2025-12-23 #9409 @boisgera
Hi @matbun,

Please,
  - have a look at https://github.com/interTwin-eu/itwinai/pull/484, then
  - perform the tasks listed in **Additional Author Tasks After Review is Complete** in [the previous comment](https://github.com/openjournals/joss-reviews/issues/9409#issuecomment-3687106638).

I'll go through the rest of my own checklist as soon as you give me the DOI of your project archive.
