### 2025-12-25 interTwin-eu/itwinai#472 @xiazeyu
> * I have added the `--ntasks` option in the slurm builder. Hopefully this should fix ["Required environment variable 'SLURM_NTASKS' not set" #481](https://github.com/interTwin-eu/itwinai/issues/481) as well

The problem is still there. I'm unsure about if `constants.py` will be executed as `itwinai exec-pipeline` seems to be a standalone command that does not involve slurm.

Note that now `itwinai run` works good, but `itwinai exec-pipeline` requires that slurm environment variable.

> *
> * regarding the `--export=ALL` flag to srun and the usage of `/bin/bash`, I would say that this is a custom requirement from your HPC, which makes totally sense. The pre-exec file is meant exactly for that -- accommodate custom setups like in your case. So in this case you can keep that script, which is customized to your HPC

I would agree with that.
