### 2025-12-25 0ce333b
Add env vars export in SLURM template
