### 2025-12-25 interTwin-eu/itwinai#472 @matbun
> > * I have added the `--ntasks` option in the slurm builder. Hopefully this should fix ["Required environment variable 'SLURM_NTASKS' not set" #481](https://github.com/interTwin-eu/itwinai/issues/481) as well
>
> The problem is still there. I'm unsure about if `constants.py` will be executed as `itwinai exec-pipeline` seems to be a standalone command that does not involve slurm.
>
> Note that now `itwinai run` works good, but `itwinai exec-pipeline` requires that slurm environment variable.

Okay, I'll reply in #481
