### 2025-12-25 interTwin-eu/itwinai#481 @matbun
Thanks @xiazeyu for reporting the full output! The problem is from this line: https://github.com/interTwin-eu/itwinai/blob/3d8a59766088e6ede3fe14a1bf78444ca896e41a/src/itwinai/distributed.py#L143

Are you in a SLURM interactive session?
