### 2025-12-25 897ba45
Fix false SLURM cluster detection in interactive session
