+++ src/itwinai/distributed.py
@@ -12,6 +12,7 @@ import builtins as __builtin__
 import functools
 import logging
 import os
+import re
 import subprocess
 import sys
 from typing import TYPE_CHECKING, Any, Callable
@@ -44,6 +45,11 @@ class ClusterEnvironment(BaseModel):
 
 
 def ray_cluster_is_running() -> bool:
+    """Check if a Ray cluster is running.
+
+    Returns:
+        bool: True if a running Ray cluster is detected. False otherwise.
+    """
     try:
         # Run the `ray status` command. It should be less overhead than ray.init()
         result = subprocess.run(
@@ -81,14 +87,18 @@ def _get_env(
     default: Any | None = None,
     cast: Callable[[str], Any] = lambda x: x,
     required: bool = False,
-) -> Any:
+) -> Any | None:
     """Fetches and casts an environment variable.
 
     Args:
-        name: the ENV var name.
-        default: returned if var is unset (and required is False).
-        cast: function to transform the raw str into the desired type.
-        required: if True and var is missing, raises KeyError.
+        name (str): the ENV var name.
+        default (Any): returned if var is unset (and required is False). Defaults to None.
+        cast (Callable[[str], Any]): function to transform the raw str into the desired type.
+            Defaults to the identity function.
+        required (bool): if True and var is missing, raises KeyError. Defaults to False.
+
+    Raises:
+        KeyError: when a required env variable is required but not found.
     """
     raw = os.getenv(name)
     if raw is None:
@@ -103,15 +113,80 @@ def _get_env(
 
 
 def _get_int(name: str, default: int | None = None, required: bool = False) -> int:
-    return _get_env(name, default=default, cast=int, required=required)
+    val = _get_env(name, default=default, cast=int, required=required)
+    if val is None:
+        raise ValueError(
+            f"Could not cast variable {name!r} to int because it is None. "
+            f"Default value was set to {default!r}"
+        )
+    return int(val)
+
+
+def _has_all(*names: str) -> bool:
+    """True iff all env vars are set (not None)."""
+    return all(os.getenv(n) is not None for n in names)
+
+
+def _is_torchrun_env() -> bool:
+    """Detect torchrun / TorchElastic cluster."""
+    # https://pytorch.org/docs/stable/elastic/run.html#environment-variables
+    return _has_all("RANK", "WORLD_SIZE", "LOCAL_RANK", "LOCAL_WORLD_SIZE") and (
+        os.getenv("TORCHELASTIC_RUN_ID") is not None or _has_all("MASTER_ADDR", "MASTER_PORT")
+    )
+
+
+def _parse_slurm_tasks_per_node(*vals: str | None) -> int | None:
+    """Parse SLURM tasks-per-node env vars.
+
+    Accepts one or more candidate strings (e.g. SLURM_NTASKS_PER_NODE,
+    SLURM_TASKS_PER_NODE). Returns the first successfully parsed value.
+
+    Handles formats like:
+      "2"
+      "2(x3)"
+      "2(x2),1"
+      "4,4,4"
+
+    Returns: max tasks on any node, or None if nothing is parseable.
+    """
+    for val in vals:
+        if not val:
+            continue
+
+        parts = [p.strip() for p in val.strip().split(",")]
+        counts: list[int] = []
+
+        for p in parts:
+            m = re.match(r"^(\d+)(?:\(x\d+\))?$", p)
+            if m:
+                counts.append(int(m.group(1)))
+
+        if counts:
+            return max(counts)
+
+    return None
 
 
[…]
