### 2025-12-25 5cf853e
Make SLURM cluster env selection stricter
