### 2025-12-25 f54a079
Remove dependency on #GPUs
