### 2025-12-25 interTwin-eu/itwinai#481 @xiazeyu
> Okay, could you try again? I have pushed a fix with more restrictive checks

I can see now this is solved.
