### 2026-01-05 efdb070
Update paper.md
