### 2026-01-05 d24751e
Fix itwinai run for MNIST (#483)
* Enhance SLURM config with comments and fields

Added detailed comments and configuration fields for SLURM job.

* Rename exp_name to experiment_name in YAML config

* Rename default pipe_key to 'training_pipeline'

* Rename default_pipe_key to 'training_pipeline'

* Add configuration parameters to run-example.yaml

* Update config_path and config_name comments

Clarified config_path and config_name in YAML.

* Add experiment and run names to YAML configuration

* Change dist_strat to use variable strategy

* Delete src/itwinai/slurm/sample_slurm_config.yaml


[…]
