+++ docs/getting-started/complete-workflow-example.rst
@@ -50,6 +50,10 @@ Steps
 
    .. code-block:: yaml
 
+      # Default fields (always needed)
+      strategy: ddp
+      run_name: "mnist"
+      
       plugins:
         - <my first required plugin>
         - <my second required plugin>
@@ -61,6 +65,27 @@ Steps
         submit_job: true   # Set this if you actually want to submit the job
         save_script: true  # Set this if you want to store the generated SLURM script(s)
         pre_exec_file: <path or URL to pre-execution file for current system>
+        
+        # Fields referring to this config file
+        pipe_key: training_pipeline
+        config_name: <name of this config file>
+        config_path: <path to the directory containing this config file>
+
+        # Propagate global config to this slurm_config
+        distributed_strategy: ${strategy}
+        run_name: ${run_name}
+
+        # This provides a template for the training command launched using:
+        # $ itwinai exec-pipeline -c <this config>.yaml [ARGS]
+        # The template is filled in using the fields of this slurm_config.
+        # Example:
+        training_cmd: >
+          {itwinai_launcher} exec-pipeline
+          --config-name={config_name}
+          --config-path={config_path}
+          --strategy={distributed_strategy}
+          --run-name={run_name}
+          +pipe_key={pipe_key}
 
         # Any other SLURM configuration options you want to set.
         # Check out the SLURM builder for more information.
@@ -87,6 +112,11 @@ Steps
 
       itwinai run -jc run_config.yaml
 
+   The ``slurm_config`` section follows :class:`itwinai.slurm.configuration.MLSlurmBuilderConfig`
+   (extending :class:`itwinai.slurm.configuration.SlurmScriptConfiguration`), which documents
+   each field and its default. Use the YAML to set values; ``-j`` and ``-s`` are the only CLI
+   overrides applied on top of the config for submission and saving.
+
 MNIST Example
 -------------
 
@@ -95,55 +125,87 @@ using the itwinai MNIST plugin. This simplified example shows the key components
 
 .. code-block:: yaml
 
-   # General training parameters
-   run_id: mnist-usecase-0
-   dataset_root: .tmp/
-   batch_size: 128
-   lr: 0.001
-   epochs: 5
-   strategy: ddp
-   # The rest of the general config is omitted for the sake of readability
+    # Default fields (always needed)
+    strategy: ddp
+    run_name: "mnist"
 
-   # Install the itwinai MNIST plugin from GitHub
-   plugins:
-     - git+https://github.com/matbun/itwinai-mnist-plugin.git
+    # General config
+    dataset_root: .tmp/
+    num_classes: 10
+    batch_size: 128
+    num_workers_dataloader: 4
+    pin_memory: False
+    lr: 0.001
+    momentum: 0.9
+    fp16_allreduce: False
+    use_adasum: False
+    gradient_predivide_factor: 1.0
+    epochs: 5
+    test_data_path: mnist-sample-data
+    inference_model_mlflow_uri: mnist-pre-trained.pth
+    predictions_dir: mnist-predictions
+    predictions_file: predictions.csv
+    class_labels: null
+    checkpoints_location: checkpoints
+    checkpoint_every: 1
 
-   # SLURM configuration for Vega HPC system
-   slurm_config:
-     job_name: mnist-job
-     account: s24r05-03-users  # Replace with your account
-     partition: gpu
-     memory: 64G
-     num_nodes: 2
-     submit_job: true
-     save_script: true
-     pre_exec_file: https://raw.githubusercontent.com/interTwin-eu/itwinai/refs/heads/main/src/itwinai/slurm/system-base-scripts/vega_pre_exec.sh
+    plugins:
+      - git+https://github.com/matbun/itwinai-mnist-plugin.git
 
[…]
