+++ tutorials/distributed-ml/torch-scaling-test/slurm_config.yaml
@@ -1,5 +1,5 @@
 # If you use this with in the runall or scaling mode, keep in mind that the strategies
-# will change, as well as the number of nodes. 
+# will change, as well as the number of nodes.
 #
 # Default arguments can be seen in src/itwinai/slurm/utils.py
 #
@@ -9,7 +9,7 @@ gpus_per_node: 4
 cpus_per_task: 16
 
 mode: single # "single", "runall" or "scaling-test" - defaults to "single"
-dist_strat: ddp # "ddp", "deepspeed" or "horovod"
+distributed_strategy: ddp # "ddp", "deepspeed" or "horovod"
 itwinai_trainer: false
 
 account: intertwin
@@ -17,21 +17,20 @@ time: 00:15:00
 partition: develbooster
 
 # Keep in mind that these will be overwritten if "mode" is not "single", and that
-# if you override the dist_strat in the CLI, then these will already have evaluated
-# and thus might not correspond. Thus, we suggest you only change the dist_strat in
+# if you override the distributed_strategy in the CLI, then these will already have evaluated
+# and thus might not correspond. Thus, we suggest you only change the distributed_strategy in
 # the config and avoid overriding it in the CLI.
-std_out: slurm_job_logs/${dist_strat}-${num_nodes}x${gpus_per_node}.out
-err_out: slurm_job_logs/${dist_strat}-${num_nodes}x${gpus_per_node}.err
-job_name: tutorial-${dist_strat}-job
+std_out: slurm_job_logs/${distributed_strategy}-${num_nodes}x${gpus_per_node}.out
+err_out: slurm_job_logs/${distributed_strategy}-${num_nodes}x${gpus_per_node}.err
+job_name: tutorial-${distributed_strategy}-job
 
 # The different number of nodes to use for the scalability testing
 scalability_nodes: "1, 2, 4"
 
 python_venv: ../../../.venv
-
 # If you want to manually override the training command, comment in the following:
-# training_cmd: | 
-#   $(which itwinai) exec-pipeline \
-#   --config_path ${config_file} \
-#   +pipe_key ${pipe_key} \
-#   strategy=${dist_strat} \
+# training_cmd: >
+#   $(which itwinai) exec-pipeline
+#   --config_path ${config_file}
+#   +pipe_key ${pipe_key}
+#   strategy=${distributed_strategy}
