+++ joss/paper.bib
@@ -72,6 +72,19 @@
   urldate   = {2025-08-14},
 }
 
+@article{manzi_intertwin_2025,
+	title = {{interTwin}: {Advancing} {Scientific} {Digital} {Twins} through {AI}, {Federated} {Computing} and {Data}},
+	issn = {0167-739X},
+	url = {https://www.sciencedirect.com/science/article/pii/S0167739X25006065},
+	doi = {https://doi.org/10.1016/j.future.2025.108312},
+	abstract = {The EU project interTwin, co-designed and implemented the prototype of an interdisciplinary Digital Twin Engine (DTE), an open-source platform that provides generic and domain-specific software components for modelling and simulation to integrate application-specific Digital Twins (DTs). The DTE is built upon a co-designed conceptual model - the DTE blueprint architecture - guided by open standards and interoperability principles. The ambition is to develop a unified approach to the implementation of DTs that is applicable across diverse scientific disciplines to foster collaborations and facilitate developments. Co-design involved DT use cases from high-energy physics, radio astronomy, astroparticle physics, climate research, and environmental monitoring, which drove advancements in modelling and simulation by leveraging heterogeneous distributed digital infrastructures, enabling dynamic workflow composition, real-time data management and processing, quality and uncertainty tracing of models, and multi-source data fusion.},
+	journal = {Future Generation Computer Systems},
+	author = {Manzi, Andrea and Bardaji, Raul and Rodero, Ivan and Moltó, Germán and Fiore, Sandro and Campos, Isabel and Elia, Donatello and Sarandrea, Francesco and Millar, A. Paul and Spiga, Daniele and Bunino, Matteo and Accarino, Gabriele and Asprea, Lorenzo and Bernardo, Samuel and Caballer, Miguel and Chatzikyriakou, Charis and Ciangottini, Diego and Claus, Michele and Cristofori, Andrea and Donno, Davide and Donno, Emanuele and Ferrario, Iacopo and Fronza, Massimiliano and Jacob, Alexander and Komijani, Javad and Marinkovic, Marina Krstic and Legger, Federica and Palomo, Ivan and Parcero, Estíbaliz and Sarma, Rakesh and Ray, Gaurav Sinha and Vallero, Sara and Zvolensky, Juraj},
+	year = {2025},
+	keywords = {Computing, Data management, Digital Twins, HPC, Machine Learning},
+	pages = {108312},
+}
+
 @article{benach_intertwin_2023,
 	title = {{interTwin} {D3}.1 {DTE} blueprint architecture, functional specifications and requirements analysis first version},
 	url = {https://zenodo.org/records/10417124},
