### 2026-01-07 23e84a8
Remove Anna's email address from author information (#486)
Removed email address for Anna Elisa Lappe.
