### 2026-01-13 492319b
JOSS submission (#485)
* First paper draft

* small fixes

* fix style checks

* Update paper.bib

* Update paper.bib

* updated plot sizes

* update images and bib

* removing repeating references

* Update paper.bib

* update refs and footnotes

* pre-incorporate rakesh bib changes

* remove trailing space

* textual changes

* Update paper.md

* small update

* add missing hpo image

* add Anna's ORCID

* Update paper.md

* add reference to Virgo exp data

* addressing comments from co-authors

* fix footnotes

* Update paper.md

* Update the paper.md with small changes

- Change the title 
[…]
