+++ joss/paper.bib
@@ -1,0 +1,268 @@
+@misc{hython,
+  author = {I. F. Ferrario and M. Alasawedah and S. Dhinakaran and M. U. Liaqat},
+  title = {Hython},
+  year = {2025},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/interTwin-eu/hython},
+  urldate = {2025-08-14},
+}
+
+@article{ray:2018,
+      title={Ray: A Distributed Framework for Emerging AI Applications}, 
+      author={Philipp Moritz and Robert Nishihara and Stephanie Wang and Alexey Tumanov and Richard Liaw and Eric Liang and Melih Elibol and Zongheng Yang and William Paul and Michael I. Jordan and Ion Stoica},
+      year={2018},
+      eprint={1712.05889},
+      archivePrefix={arXiv},
+      primaryClass={cs.DC},
+      url={https://arxiv.org/abs/1712.05889}, 
+}
+
+@misc{interlink,
+	title = {interLink},
+	author = {D. Ciangottini and G. Bianchini and D. Spiga},
+	copyright = {Apache-2.0},
+	url = {https://github.com/interLink-hq/interLink},
+	abstract = {InterLink aims to provide an abstraction for the execution of a Kubernetes pod on any remote resource capable of managing a Container execution lifecycle},
+	urldate = {2025-07-27},
+	publisher = {GitHub},
+	journal = {GitHub repository},
+	month = feb,
+	year = {2025},
+}
+
+@inproceedings{optuna_2019,
+  author    = {Akiba, Takuya and Sano, Shotaro and Yanase, Toshihiko and Ohta, Takeru and Koyama, Masanori},
+  title     = {Optuna: A Next-generation Hyperparameter Optimization Framework},
+  booktitle = {Proceedings of the 25th {ACM} {SIGKDD} International Conference on Knowledge Discovery and Data Mining},
+  year      = {2019},
+  pages     = {2623--2631},
+  publisher = {ACM},
+  doi       = {10.1145/3292500.3330701},
+  url       = {https://doi.org/10.1145/3292500.3330701},
+}
+
+@misc{heat,
+  author    = {HeAT Developers},
+  title     = {HeAT},
+  year      = {2025},
+  publisher = {GitHub},
+  journal   = {GitHub repository},
+  url       = {https://github.com/helmholtz-analytics/heat},
+  urldate   = {2025-08-14},
+}
+
+@misc{ai4hpc,
+  author    = {AI4HPC Developers},
+  title     = {AI4HPC},
+  year      = {2025},
+  publisher = {GitLab},
+  journal   = {GitLab repository},
+  url       = {https://gitlab.jsc.fz-juelich.de/CoE-RAISE/FZJ/ai4hpc/ai4hpc},
+  urldate   = {2025-08-14},
+}
+
+@misc{perun,
+  author    = {Helmholtz AI Energy Team},
+  title     = {perun},
+  year      = {2025},
+  publisher = {GitHub},
+  journal   = {GitHub repository},
+  url       = {https://github.com/Helmholtz-AI-Energy/perun},
+  urldate   = {2025-08-14},
+}
+
+@article{manzi_intertwin_2025,
+	title = {{interTwin}: {Advancing} {Scientific} {Digital} {Twins} through {AI}, {Federated} {Computing} and {Data}},
+	issn = {0167-739X},
+	url = {https://www.sciencedirect.com/science/article/pii/S0167739X25006065},
+	doi = {https://doi.org/10.1016/j.future.2025.108312},
[…]
