+++ joss/paper.md
@@ -1,0 +1,290 @@
+---
+title: "itwinai: A Python Toolkit for Scalable Scientific Machine Learning on HPC Systems"
+tags:
+  - Python
+  - Digital Twins
+  - Distributed Training
+  - Hyperparameter Optimization
+  - High Performance Computing
+authors:
+  - given-names: Matteo
+    surname: Bunino
+    orcid: 0009-0008-5100-9300
+    affiliation: 1
+    corresponding: true 
+  - given-names: Jarl Sondre
+    surname: Sæther
+    orcid: 0009-0002-7971-2213
+    affiliation: 1
+  - given-names: Linus Maximilian
+    surname: Eickhoff
+    orcid: 0009-0006-6691-2821
+    affiliation: 1
+  - given-names: Anna Elisa
+    surname: Lappe
+    affiliation: 1
+    orcid: 0009-0009-4804-4188
+  - given-names: Kalliopi
+    surname: Tsolaki
+    email: kalliopi.tsolaki@cern.ch
+    affiliation: 1
+    orcid: 0000-0002-3192-4260
+  - given-names: Killian
+    surname: Verder
+    email: killian.verder@cern.ch
+    affiliation: 1
+    orcid: 0009-0006-4819-3229
+  - given-names: Henry
+    surname: Mutegeki
+    email: henry.mutegeki@cern.ch
+    affiliation: 1
+    orcid: 0009-0001-9940-1167
+  - given-names: Roman
+    surname: Machacek
+    email: roman.machacek@cern.ch
+    affiliation: 1
+    orcid: 0009-0007-9976-4420
+  - given-names: Maria
+    surname: Girone
+    affiliation: 1
+    orcid: 0000-0003-0261-8392
+  - given-names: Oleksandr
+    surname: Krochak
+    orcid: 0009-0007-2245-9452
+    affiliation: 2
+  - given-names: Mario
+    surname: Rüttgers
+    orcid: 0000-0003-3917-8407
+    affiliation: "2, 3"
+  - given-names: Rakesh
+    surname: Sarma
+    orcid: 0000-0002-7069-4082
+    affiliation: 2
+  - given-names: Andreas
+    surname: Lintermann
+    orcid: 0000-0003-3321-6599
+    affiliation: 2
+affiliations:
+ - name: European Organization for Nuclear Research (CERN), Espl. des Particules 1, 1217 Genève, Switzerland
+   index: 1
+ - name: Forschungszentrum Jülich, Jülich Supercomputing Center, Germany
+   index: 2
+ - name: Data-Driven Fluid Engineering (DDFE) Laboratory, Inha University, Incheon, Republic of Korea
+   index: 3
+date: 16 August 2025
+bibliography: paper.bib
+
+---
+
+<!-- markdownlint-disable MD025 -->
+
+# Summary
+
+The integration of Artificial Intelligence (AI) into scientific research has
+expanded significantly over the past decade, driven by the availability of
+large-scale datasets and Graphics Processing Units (GPUs),
+in particular at High Performance Computing (HPC) sites.
+
+However, many researchers face significant barriers
+when deploying AI workflows on HPC systems, as their heterogeneous nature forces
+scientists to focus on low-level implementation details rather than on their core
+research. At the same time, the researchers often lack specialized HPC/AI knowledge
+to implement their workflows efficiently.
+
+To address this, we present `itwinai`, a Python library that simplifies scalable
+AI on HPC. Its modular architecture and standard interface allow users to scale
+workloads efficiently from laptops to supercomputers, reducing implementation
+overhead and improving resource usage.
+
+# Statement of need
+
+Integrating machine learning into scientific workflows on HPC systems remains
+complex. Researchers must often invest substantial effort to configure
[…]
