### 2026-01-13 8004e36
Bump version and disable HPC tests (no compute time available atm)
