### 2026-01-14 f1ec625
adding cli commands for pushing to ai model hub
