+++ src/itwinai/cli.py
@@ -8,6 +8,7 @@
 # - Jarl Sondre Sæther <jarl.sondre.saether@cern.ch> - CERN
 # - Anna Lappe <anna.elisa.lappe@cern.ch> - CERN
 # - Linus Eickhoff <linus.maximilian.eickhoff@cern.ch> - CERN
+# - Rakesh Sarma <r.sarma@fz-juelich.de> - FZJ
 #
 # --------------------------------------------------------------------------------------
 # Command-line interface for the itwinai Python library.
@@ -37,6 +38,9 @@ from typing_extensions import Annotated
 from validators import url
 
 from itwinai.constants import BASE_EXP_NAME, RELATIVE_MLFLOW_PATH
+
+import requests
+import certifi
 
 app = typer.Typer(pretty_exceptions_enable=False)
 
@@ -948,6 +952,209 @@ def tensorboard_ui(
 
     subprocess.run(f"tensorboard --logdir={path} --port={port} --host={host}".split())
 
+
+@app.command()
+def upload_model_to_hub(
+    model_dir: Annotated[
+        str,
+        typer.Argument(
+            help="Path to directory with model checkpoint, manifest.yaml and metadata."
+        ),
+    ],
+    hub_url: Annotated[
+        str | None,
+        typer.Option(
+            "--hub-url",
+            help="URL of model hub server. If not provided, use HYPHA_SERVER_URL or .env file",
+        ),
+    ] = None,
+    api_token: Annotated[
+        str | None,
+        typer.Option(
+            "--api-token",
+            help="API token. If not provided, use HYPHA_TOKEN or .env file.",
+        ),
+    ] = None,
+    env_file: Annotated[
+        str | None,
+        typer.Option(
+            "--env-file",
+            help="Path to .env file containing MODEL_HUB_URL and MODEL_HUB_API_TOKEN.",
+        ),
+    ] = None,
+    upload_script: Annotated[
+        str | None,
+        typer.Option(
+            "--upload-script",
+            help="Path to upload_model.py script. If not provided, downloads from GitHub.",
+        ),
+    ] = None,
+):
+    """Upload a model checkpoint to the AI Model Hub.
+    Please note that this command requires internet connection to push to the model hub.
+
+    The model directory should contain:
+    - model checkpoint file(s)
+    - manifest.yaml with model id and metadata
+    - (optional) metadata.json with additional information
+    """
+    import subprocess
+    import tempfile
+
+    model_path = Path(model_dir).resolve()
+
+    # Validate if model directory exists and is a directory
+    if not model_path.exists():
+        cli_logger.error(f"Model directory '{model_path}' does not exist!")
+        raise typer.Exit(code=1)
+
+    if not model_path.is_dir():
+        cli_logger.error(f"'{model_path}' is not a directory!")
+        raise typer.Exit(code=1)
+
+    # Check if the file manifest.yaml exists
+    manifest_file = model_path / "manifest.yaml"
+    if not manifest_file.exists():
+        cli_logger.error(
+            f"No manifest.yaml found in '{model_path}'. "
+            "The model directory must contain a manifest.yaml file with the model id."
+        )
+        raise typer.Exit(code=1)
+
+    # Load environment variables from .env file if specified and if file exists
+    env_vars = os.environ.copy()
+
+    if env_file:
+        env_path = Path(env_file)
+        if not env_path.exists():
+            cli_logger.error(f"Specified .env file '{env_path}' does not exist!")
+            raise typer.Exit(code=1)
+        _load_env_file(env_path, env_vars)
+        cli_logger.info(f"Loaded environment from {env_path}")
+    elif Path(".env").exists():
+        _load_env_file(Path(".env"), env_vars)
[…]
