### 2026-01-14 #9409 @matbun
> > > Hi [@matbun](https://github.com/matbun),
> > > Please have a look at:
> > >
> > > * [Fix DOI format in paper.bib entry interTwin-eu/itwinai#491](https://github.com/interTwin-eu/itwinai/pull/491)
> >
> >
> > I am going to merge this on main soon. Should I also create a new software release including the fixed URL or the last one I made is OK? [interTwin-eu/itwinai#492](https://github.com/interTwin-eu/itwinai/pull/492)
>
> I don't think that this really minor issue requires a bug release; I'll just rerun the check once you've merged the PR in main. Let's proceed with your current release; if I am mistaken, the editor in chief which has the last word on the paper acceptance will tell us.

I have merged the new PR
