### 2026-01-14 #9409 @matbun
> Hi [@matbun](https://github.com/matbun),
>
> The pdf generation currently fails with `Couldn't check the bibtex because branch name is incorrect: joss (Theoj::Error)`. I (wild) guess that there is a mismatch the branch your bib file is and the branch that was declared to contain the joss paper?

@boisgera , yes, the `joss` branch doesnt exist anymore as I deleted it after merging the last fix. Now `main` is up-to-date and we could use `main`. Is it OK for you? In the meantime, I have restored the `joss` branch to have the build pass
