### 2026-01-19 #9409 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
