### 2026-01-20 #9409 @crvernon
@matbun - I just need you to address the following before I accept this submission for publication.

In your paper:

- [x] LINE 115:  "...interferometer (Tsolaki et al., 2025) (Sæther et al., 2025)." should be cited in-text like: "...interferometer (Tsolaki et al., 2025; Sæther et al., 2025)."

Let me know when this change has been made.
