### 2026-01-20 #9409 @matbun
@crvernon the citation format should be fixed now :)
